% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{add_interaction_type}
\alias{add_interaction_type}
\title{add_interaction_type}
\usage{
add_interaction_type(network)
}
\arguments{
\item{network}{a network table/list}
}
\value{
a network table/list with network_add interaction type
}
\description{
Add interaction type to static or dynamic network
}
