% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregating.data}
\alias{aggregating.data}
\title{Create aggregated data for inferCSN}
\usage{
aggregating.data(
  object,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction_name = NULL,
  size_factor_normalize = TRUE,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{object}{An expression matrix, cells by genes.}

\item{k_neigh}{Number of cells to be aggregated per cluster.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary}

\item{max_overlap}{The maximum overlapping ratio of two clusters.}

\item{reduction_name}{The reduction name of extracting the cell coordinates used for aggregating.}

\item{size_factor_normalize}{Logical, should accessibility values be normalized by size factor.}

\item{seed}{The seed used in randomly shuffling the data for cross-validation.}

\item{verbose}{Print detailed information.}
}
\value{
Seurat object with aggregated data
}
\description{
Function to generate aggregated inputs for inferCSN. \code{aggregating.data}
}
