% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregating_data}
\alias{aggregating_data}
\title{Create aggregated data for inferCSN}
\usage{
aggregating_data(
  object,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction_name = NULL,
  size_factor_normalize = TRUE,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{object}{The input data for \emph{\code{inferCSN}}.}

\item{k_neigh}{Number of cells to be aggregated per cluster.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary}

\item{max_overlap}{The maximum overlapping ratio of two clusters.}

\item{reduction_name}{The reduction name of extracting the cell coordinates used for aggregating.}

\item{size_factor_normalize}{Logical, should accessibility values be normalized by size factor.}

\item{seed}{The random seed for cross-validation, default is \emph{\code{1}}.}

\item{verbose}{Logical value, default is \emph{\code{TRUE}}, whether to print progress messages.}
}
\value{
Seurat object with aggregated data
}
\description{
Function to generate aggregated inputs for inferCSN. \code{aggregating_data}
}
