% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{as.slam.matrix}
\alias{as.slam.matrix}
\title{Convert a sparseMatrix from Matrix package to a slam matrix}
\usage{
as.slam.matrix(sp_mat)
}
\arguments{
\item{sp_mat}{The matrix for the aggregated single cell data}
}
\description{
Convert a sparseMatrix from Matrix package to a slam matrix
}
