% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{as.sparse}
\alias{as.sparse}
\title{Convert a slam matrix to a sparseMatrix}
\usage{
as.sparse(slam_matrix)
}
\arguments{
\item{slam_matrix}{A slam matrix}
}
\description{
Convert a slam matrix to a sparseMatrix
}
