% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{as_sparse}
\alias{as_sparse}
\title{Convert a slam matrix to a sparseMatrix}
\usage{
as_sparse(slam_matrix)
}
\arguments{
\item{slam_matrix}{A slam matrix}
}
\description{
Convert a slam matrix to a sparseMatrix
}
