% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.network.R
\name{assign_epochs_simple}
\alias{assign_epochs_simple}
\title{Assigns genes to epochs just based on which mean is maximal}
\usage{
assign_epochs_simple(
  matrix,
  dynamic_object,
  num_epochs = 3,
  pThresh = 0.01,
  toScale = FALSE,
  key_word = "epoch"
)
}
\arguments{
\item{matrix}{expression matrix}

\item{dynamic_object}{result of running findDynGenes}

\item{num_epochs}{num_epochs}

\item{pThresh}{pThresh}

\item{toScale}{toScale}

\item{key_word}{key_word}
}
\value{
data.frame of dynamically expressed genes, cluster, peakTime, ordered by peaktime
}
\description{
Assigns genes to epochs just based on which mean is maximal
}
