% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_degree_distribution}
\alias{calculate_degree_distribution}
\alias{calculate_degree_distribution,Network-method}
\alias{calculate_degree_distribution,data.frame-method}
\title{calculate_degree_distribution}
\usage{
calculate_degree_distribution(object, ...)

\S4method{calculate_degree_distribution}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = TRUE,
  ...
)

\S4method{calculate_degree_distribution}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = TRUE,
  ...
)
}
\arguments{
\item{object}{The input object.}

\item{...}{Parameters for other methods.}

\item{regulators}{regulators}

\item{targets}{targets}

\item{directed}{directed}
}
\description{
calculate_degree_distribution
}
