% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_gene_rank}
\alias{calculate_gene_rank}
\alias{calculate_gene_rank,Network-method}
\alias{calculate_gene_rank,data.frame-method}
\title{Calculate gene ranks in network}
\usage{
calculate_gene_rank(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)

\S4method{calculate_gene_rank}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)

\S4method{calculate_gene_rank}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{object}{A Network object or data frame containing network information}

\item{regulators}{Optional character vector of regulator genes}

\item{targets}{Optional character vector of target genes}

\item{directed}{Logical, whether the network is directed}
}
\value{
A data frame of gene ranks
}
\description{
Calculate gene importance using PageRank algorithm
}
