% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_page_rank}
\alias{calculate_page_rank}
\alias{calculate_page_rank,Network-method}
\alias{calculate_page_rank,data.frame-method}
\title{Calculate PageRank}
\usage{
calculate_page_rank(object, ...)

\S4method{calculate_page_rank}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)

\S4method{calculate_page_rank}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{object}{Network object}

\item{...}{Additional arguments}

\item{regulators}{Character vector, regulators to include}

\item{targets}{Character vector, targets to include}

\item{directed}{Logical, whether the network is directed}
}
\value{
Data frame with PageRank scores
}
\description{
Calculate PageRank
}
