% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{caoGenes}
\alias{caoGenes}
\title{cluster and order genes}
\usage{
caoGenes(
  expSmooth,
  matrix,
  dynamic_object,
  k = 3,
  pThresh = 0.01,
  method = "kmeans"
)
}
\arguments{
\item{expSmooth}{expression matrix}

\item{matrix}{matrix}

\item{dynamic_object}{result of running findDynGenes}

\item{k}{number of clusters}

\item{pThresh}{pval threshold}

\item{method}{method}
}
\value{
data.frame of dynamically expressed genes, cluster, peakTime, ordered by peaktime
}
\description{
cluster and order genes
}
