% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.functions.R
\name{celltype.annotation}
\alias{celltype.annotation}
\title{annotation.celltype}
\usage{
celltype.annotation(object, method = "celltypist", reference_path = NULL)
}
\arguments{
\item{object}{seurat object}

\item{method}{}
}
\value{
seurat object
}
\description{
annotation.celltype
}
