% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.parameters}
\alias{check.parameters}
\title{Check input parameters}
\usage{
check.parameters(
  matrix,
  penalty,
  algorithm,
  crossValidation,
  seed,
  nFolds,
  kFolds,
  rThreshold,
  regulators,
  targets,
  maxSuppSize,
  verbose,
  cores
)
}
\value{
No return value, called for check input parameters
}
\description{
Check input parameters
}
