% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{coef.Network}
\alias{coef.Network}
\title{Get fitted coefficients}
\usage{
\method{coef}{Network}(object, ...)
}
\arguments{
\item{object}{Network object}

\item{...}{Other parameters}
}
\value{
Return the fitted coefficients
}
\description{
Get fitted coefficients
}
