% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{coef,CSNObject-method}
\alias{coef,CSNObject-method}
\alias{coef,Network-method}
\title{Get fitted coefficients}
\usage{
\S4method{coef}{CSNObject}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\S4method{coef}{Network}(object, ...)
}
\arguments{
\item{object}{The csn object}

\item{network}{network}

\item{celltypes}{cell types to analyze, NULL for all cell types}

\item{...}{other parameters}
}
\description{
Get fitted coefficients
}
