% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{compute_betweenness_degree}
\alias{compute_betweenness_degree}
\title{Function to compute betweenness and degree}
\usage{
compute_betweenness_degree(
  networks_list,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{networks_list}{result of dynamic network reconstruction}

\item{regulators}{list of regulators}

\item{targets}{list of targets}

\item{directed}{if network is directed or not}
}
\value{
list of each network, ranked by `betweenness*degree`
}
\description{
Function to compute betweenness and degree
}
