% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{compute_pagerank}
\alias{compute_pagerank}
\title{Function to compute page rank of TF+target networks}
\usage{
compute_pagerank(
  networks_list,
  regulators = NULL,
  targets = NULL,
  directed = FALSE
)
}
\arguments{
\item{networks_list}{result of dynamic network reconstruction}

\item{regulators}{list of regulators}

\item{targets}{list of targets}

\item{directed}{if network is directed or not}
}
\value{
list of each network, ranked by `page_rank`
}
\description{
Function to compute page rank of TF+target networks
}
