% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.analysis.R
\name{cor_and_add_action}
\alias{cor_and_add_action}
\title{Adds an extra column to the result of dynamic_shortest_path_multiple that predicts overall action based on correlation between "from" and "to"}
\usage{
cor_and_add_action(spDF, matrix)
}
\arguments{
\item{spDF}{the result of running dynamic_shortest_path_multiple}

\item{matrix}{corresponding genes-by-cells expression matrix}
}
\value{
shortest paths dataframe with added action_by_corr column
}
\description{
Adds an extra column to the result of dynamic_shortest_path_multiple that predicts overall action based on correlation between "from" and "to"
}
