% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{CSNObject-class}
\alias{CSNObject-class}
\alias{CSNObject}
\title{The CSNObject class}
\description{
The CSNObject object is an extended \code{Seurat} object
for the storage and analysis of Regulatory network data.
}
\section{Slots}{

\describe{
\item{\code{grn}}{A named list containing \code{RegulatoryNetwork} objects with inferred networks.}

\item{\code{data}}{Seurat object.}
}}

