% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{CSNObjectList-class}
\alias{CSNObjectList-class}
\alias{CSNObjectList}
\title{The CSNObjectList class}
\description{
The CSNObjectList object is an extended \code{Seurat} object
for the storage and analysis of Regulatory network data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Seurat object list.}
}}

