% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{DefaultNetwork}
\alias{DefaultNetwork}
\alias{DefaultNetwork.CSNObject}
\alias{DefaultNetwork.RegulatoryNetwork}
\title{Get active network}
\usage{
DefaultNetwork(object, ...)

\method{DefaultNetwork}{CSNObject}(object, ...)

\method{DefaultNetwork}{RegulatoryNetwork}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get active network
}
