% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_intersection.R
\name{density_points}
\alias{density_points}
\title{density_points}
\usage{
density_points(
  meta_data,
  cluster_list = NULL,
  method = "combine",
  celltype_by = NULL,
  group_column = "cluster",
  pseudotime_column = "pseudotime",
  min_cells = 100,
  plot = FALSE,
  key_word = "network_",
  key_sign = "..",
  file_save = NULL,
  color_list = c("#0066ff", "#0099ff", "#66cc33", "#66ff33", "#ccff00", "#ffff99",
    "#ffff33", "#ffcc33", "#ff9933", "#ff6633", "#cc3300", "#ff3333")
)
}
\arguments{
\item{meta_data}{meta_data}

\item{cluster_list}{cluster_list}

\item{method}{method}

\item{celltype_by}{celltype_by}

\item{group_column}{group_column}

\item{pseudotime_column}{pseudotime_column}

\item{min_cells}{min_cells}

\item{plot}{plot}

\item{key_word}{key_word}

\item{key_sign}{key_sign}

\item{file_save}{file_save}

\item{color_list}{color_list}
}
\value{
points
}
\description{
density_points
}
\examples{
test_data <- rbind(
  data.frame(
    cluster = "cluster1",
    pseudotime = rnorm(500, mean = 1, sd = 1)
  ),
  data.frame(
    cluster = "cluster2",
    pseudotime = rnorm(500, mean = 2, sd = 2)
  ),
  data.frame(
    cluster = "cluster3",
    pseudotime = rnorm(500, mean = 3, sd = 3)
  )
)

density_points(
  meta_data = test_data,
  plot = TRUE
)

density_points(
  meta_data = test_data,
  cluster_list = list(c("cluster1", "cluster3")),
  plot = TRUE,
  color_list = c("#ff0000", "#ffcc00")
)

density_points(
  meta_data = test_data,
  method = "intersection",
  plot = TRUE
)
}
