% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-utils.R
\name{dMcast}
\alias{dMcast}
\title{Copy of the dMcast function from the Matrix.utils package,
since this is off CRAN and does not seem to be maintained anymore}
\usage{
dMcast(
  data,
  formula,
  fun.aggregate = "sum",
  value.var = NULL,
  as.factors = FALSE,
  factor.nas = TRUE,
  drop.unused.levels = TRUE
)
}
\arguments{
\item{data}{data}

\item{formula}{formula}

\item{fun.aggregate}{fun.aggregate}

\item{value.var}{value.var}

\item{as.factors}{as.factors}

\item{factor.nas}{factor.nas}

\item{drop.unused.levels}{drop.unused.levels}
}
\value{
data
}
\description{
Copy of the dMcast function from the Matrix.utils package,
since this is off CRAN and does not seem to be maintained anymore
}
