% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod-dynamic.genes.R
\name{dynamic.genes_new}
\alias{dynamic.genes_new}
\title{find genes expressed dynamically}
\usage{
dynamic.genes_new(
  object,
  meta_data,
  cluster_by = NULL,
  group_column = "cluster",
  pseudotime_column = "pseudotime",
  min_cells = 100,
  p_value = 0.05,
  verbose = TRUE,
  cores = 1
)
}
\arguments{
\item{object}{properly normalized expression matrix}

\item{meta_data}{sample table that includes pseudotime, rownames = cells, and a group column}

\item{cluster_by}{vector of group names to include}

\item{group_column}{column name in meta_data annotating groups in the cluster_by}

\item{pseudotime_column}{column name in meta_data annotating pseudotime or latent time}

\item{min_cells}{min cells}

\item{p_value}{p_value = 0.05}

\item{verbose}{verbose}

\item{cores}{cores}
}
\value{
pvals and cell info
}
\description{
find genes expressed dynamically
}
