% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod-dynamic_genes.R
\name{dynamic_genes}
\alias{dynamic_genes}
\alias{dynamic_genes.default}
\alias{dynamic_genes.Seurat}
\title{Get dynamic genes}
\usage{
dynamic_genes(object, ...)

\method{dynamic_genes}{default}(object, pseudotime = NULL, fdr_threshold = 0.05, cores = 1, ...)

\method{dynamic_genes}{Seurat}(
  object,
  fdr_threshold = 0.05,
  cluster = NULL,
  cluster_column = NULL,
  pseudotime_column = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{object}{Expression object.}

\item{...}{Arguments for other methods}

\item{pseudotime}{Pseudotime}

\item{fdr_threshold}{Threshold of fdr.}

\item{cores}{cores}

\item{cluster}{cluster}

\item{cluster_column}{idents}

\item{pseudotime_column}{pseudotime_column}

\item{slot}{slot used}
}
\value{
A new object with dynamic genes

Gene list
}
\description{
Get dynamic genes
}
\examples{
\dontrun{
data("example_matrix")
vector_result <- infer_vector(example_matrix)
dynamic_genes(
  object = t(vector_result$matrix),
  pseudotime = vector_result$pseudotime[, 2]
)
}
}
