% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{dynamic_shortest_path_multiple}
\alias{dynamic_shortest_path_multiple}
\title{Function to return shortest path from multiple TFs to multiple targets in a dynamic network}
\usage{
dynamic_shortest_path_multiple(
  network_table,
  regulators,
  targets,
  weight_column = "weight"
)
}
\arguments{
\item{network_table}{a dyanmic network}

\item{regulators}{the starting TFs}

\item{targets}{the end TFs}

\item{weight_column}{column name in network_table with edge weights that will be converted to distances}
}
\value{
dataframe with shortest path, distance, normalized distance, and action
}
\description{
Function to return shortest path from multiple TFs to multiple targets in a dynamic network
}
