% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_intersection.R
\name{dynamic_windowing}
\alias{dynamic_windowing}
\title{dynamic_windowing}
\usage{
dynamic_windowing(
  meta_data,
  bin_points = NULL,
  cluster_list = NULL,
  celltype_by = NULL,
  group_column = "cluster",
  pseudotime_column = "pseudotime",
  min_cells = 100,
  key_word = "network_",
  key_sign = ".."
)
}
\arguments{
\item{meta_data}{meta_data}

\item{bin_points}{bin_points}

\item{cluster_list}{cluster_list}

\item{celltype_by}{celltype_by}

\item{group_column}{group_column}

\item{pseudotime_column}{pseudotime_column}

\item{min_cells}{min_cells}

\item{key_word}{key_word}

\item{key_sign}{key_sign}
}
\value{
Divided cell list
}
\description{
dynamic_windowing
}
