% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{export_csn}
\alias{export_csn}
\alias{export_csn,CSNObject-method}
\alias{export_csn,Network-method}
\title{Export network from CSN object}
\usage{
export_csn(object, ...)

\S4method{export_csn}{CSNObject}(
  object,
  active_network = NULL,
  celltypes = NULL,
  weight_cutoff = NULL,
  ...
)

\S4method{export_csn}{Network}(object, weight_cutoff = NULL, ...)
}
\arguments{
\item{object}{A CSNObject object}

\item{...}{Additional arguments}

\item{active_network}{Character string specifying which network to export}

\item{celltypes}{Character vector of cell types to export}

\item{weight_cutoff}{Numeric threshold for filtering edges by absolute weight}
}
\value{
A list of network data frames by cell type
}
\description{
Export network data from a CSN object with optional filtering
}
