% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.network.R
\name{extract.network}
\alias{extract.network}
\title{extract.network}
\usage{
extract.network(object, cluster)
}
\arguments{
\item{object}{Seurat object with network results after run 'inferCSN()'}

\item{cluster}{Selected celltype(s)}
}
\value{
network
}
\description{
extract.network
}
