% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.network.R
\name{extract.nodes}
\alias{extract.nodes}
\title{extract Node character}
\usage{
extract.nodes(network_links, markers)
}
\arguments{
\item{network_links}{Links obtained by extract.links}

\item{markers}{Dataframe with marker and cluster information}
}
\value{
Node character dataframe
}
\description{
extract Node character
}
