% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network.R
\name{extract_cre.gene.links}
\alias{extract_cre.gene.links}
\title{Extract cre-gene relations of markers}
\usage{
extract_cre.gene.links(network_table, markers, filter = FALSE)
}
\arguments{
\item{network_table}{The result of \code{inferCSN()}}

\item{markers}{A dataframe data with marker genes and cluster information,
such as the results of \code{Seurat::FindAllMarkers(object)}.}

\item{filter}{filter}
}
\description{
Extract cre-gene relations of markers
}
