% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network.R
\name{extract_cre}
\alias{extract_cre}
\title{extract cre of markers}
\usage{
extract_cre(cre_gene_list, promoter_cre_gene_list, da_peaks_list = NULL)
}
\arguments{
\item{cre_gene_list}{a list of CRE-Gene relationships}

\item{promoter_cre_gene_list}{a list of Promoter-Gene relationships}

\item{da_peaks_list}{a list of DA of each cluster}
}
\description{
extract cre of markers
}
