% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network.R
\name{extract_network}
\alias{extract_network}
\title{extract_network}
\usage{
extract_network(object, cluster)
}
\arguments{
\item{object}{Seurat object with network results after run 'inferCSN()'}

\item{cluster}{Selected celltype(s)}
}
\value{
network
}
\description{
extract_network
}
