% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network.R
\name{extract_nodes}
\alias{extract_nodes}
\title{extract Node character}
\usage{
extract_nodes(network_links, markers)
}
\arguments{
\item{network_links}{Links obtained by extract_links}

\item{markers}{Dataframe with marker and cluster information}
}
\value{
Node character dataframe
}
\description{
extract Node character
}
