% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{final.network}
\alias{final.network}
\title{final.network}
\usage{
final.network(object, cluster)
}
\arguments{
\item{object}{Seurat object with network results after run 'inferCSN()'}

\item{cluster}{Selected celltype(s)}
}
\value{
network
}
\description{
final.network
}
