% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{find_communities}
\alias{find_communities}
\title{find_communities}
\usage{
find_communities(network_table, use_weights = FALSE, directed = FALSE)
}
\arguments{
\item{network_table}{network_table}

\item{use_weights}{whether or not to use edge weights (for weighted graphs)}

\item{directed}{Default set to `FALSE`.
If directed = TRUE, remember to flip target and regulator in diffnet dfs}
}
\value{
community assignments of nodes in the dynamic network
}
\description{
Find communities in a static or dynamic network
}
