% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.network.R
\name{find_cuts_by_similarity}
\alias{find_cuts_by_similarity}
\title{find_cuts_by_similarity}
\usage{
find_cuts_by_similarity(
  matrix,
  dynamic_object,
  winSize = 2,
  limit_to = NULL,
  p_value = 0.05
)
}
\arguments{
\item{matrix}{genes-by-cells expression matrix}

\item{dynamic_object}{result of running findDynGenes or define_epochs}

\item{winSize}{number of cells to each side to compare each cell to}

\item{limit_to}{vector of genes on which to base epoch cuts, for example, limiting to TFs}

\item{p_value}{pval threshold if gene is dynamically expressed}
}
\value{
vector of  pseudotimes at which to cut data into epochs
}
\description{
Returns cuts to define epochs via sliding window comparison
}
