% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod-dynamic_genes.R
\name{findDynGenes}
\alias{findDynGenes}
\title{find genes expressed dynamically}
\usage{
findDynGenes(
  object,
  meta_data,
  cluster_by = NULL,
  group_column = "cluster",
  pseudotime_column = "pseudotime",
  cores = 1
)
}
\arguments{
\item{object}{properly normalized expression matrix}

\item{meta_data}{sample table that includes pseudotime, rownames = cells, and a group column}

\item{cluster_by}{vector of group names to include}

\item{group_column}{column name in meta_data annotating groups in the cluster_by}

\item{pseudotime_column}{column name in meta_data annotating pseudotime or latent time}

\item{cores}{cores}
}
\value{
pvals and cell info
}
\description{
find genes expressed dynamically
}
