% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_brms}
\alias{fit_brms}
\title{Fit a Bayesian regression model with brms and Stan}
\usage{
fit_brms(
  formula,
  data,
  family = gaussian,
  prior = brms::prior(normal(0, 1)),
  ...
)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{family}{A description of the error distribution and link function.
See \emph{\code{stats::family}} for details.}

\item{prior}{The prior distribution specification for coefficients.
Default \emph{\code{normal(0,1)}} provides ridge-like regularization.
See \emph{\code{brms::set_prior}} for details.}

\item{...}{Additional parameters passed to \emph{\code{brms::brm}}.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{gof}} - Bayesian R-squared and model fit metrics
\item \emph{\code{coefs}} - Posterior estimates with uncertainty intervals
}
}
\description{
Fits a Bayesian regression model using the brms interface to Stan.
}
