% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit generalized linear model}
\usage{
fit_glm(formula, data, family = gaussian, ...)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{family}{A description of the error distribution and link function.
See \emph{\code{stats::family}} for details.}

\item{...}{Additional parameters passed to \emph{\code{stats::glm}}.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{metrics}} - Goodness of fit measures including R-squared
\item \emph{\code{coefficients}} - Fitted coefficients with standard errors and p-values
}
}
\description{
Fits a generalized linear model using standard maximum likelihood estimation.
}
