% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit regularized generalized linear model}
\usage{
fit_glmnet(formula, data, family = gaussian, alpha = 0.5, ...)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{family}{A description of the error distribution and link function.
See \emph{\code{stats::family}} for details.}

\item{alpha}{The elasticnet mixing parameter, between 0 and 1:
\itemize{
\item \emph{\code{alpha = 1}}: Lasso penalty
\item \emph{\code{alpha = 0}}: Ridge penalty
\item \emph{\verb{0 < alpha < 1}}: Elastic net penalty
}}

\item{...}{Additional parameters passed to \emph{\code{glmnet::glmnet}}.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{metrics}} - Goodness of fit measures including lambda and R-squared
\item \emph{\code{coefficients}} - Regularized coefficient estimates
}
}
\description{
Fits a regularized GLM using elastic net penalties via glmnet.
}
