% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_inferCSN.R
\name{fit_grn_models}
\alias{fit_grn_models}
\alias{fit_grn_models.CSNObject}
\title{Fit models for gene expression}
\usage{
fit_grn_models(object, ...)

\method{fit_grn_models}{CSNObject}(
  object,
  genes = NULL,
  network_name = paste0(method, "_network"),
  peak_to_gene_method = c("Signac", "GREAT"),
  upstream = 1e+05,
  downstream = 0,
  extend = 1e+06,
  only_tss = FALSE,
  peak_to_gene_domains = NULL,
  parallel = FALSE,
  tf_cor = 0.1,
  peak_cor = 0,
  aggregate_rna_col = NULL,
  aggregate_peaks_col = NULL,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  interaction_term = ":",
  adjust_method = "fdr",
  scale = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}

\item{genes}{A character vector with the target genes to consider for GRN inference.
Takes all VariableFeatures in the object per default.}

\item{network_name}{network_name}

\item{peak_to_gene_method}{Character specifying the method to
link peak overlapping motif regions to nearby genes. One of 'Signac' or 'GREAT'.}

\item{upstream}{Integer defining the distance upstream of the gene to consider as potential regulatory region.}

\item{downstream}{Integer defining the distance downstream of the gene to consider as potential regulatory region.}

\item{extend}{Integer defining the distance from the upstream and downstream of the basal regulatory region.
Only used of `peak_to_gene_method = 'GREAT'`.}

\item{only_tss}{Logical. Measure distance from the TSS (\code{TRUE}) or from the entire gene body (\code{FALSE}).}

\item{peak_to_gene_domains}{`GenomicRanges` object with regulatory regions for each gene.}

\item{parallel}{Logical. Whether to parallelize the computation with \code{\link[foreach]{foreach}}.}

\item{tf_cor}{Threshold for TF - target gene correlation.}

\item{peak_cor}{Threshold for binding peak - target gene correlation.}

\item{aggregate_rna_col}{aggregate_rna_col}

\item{aggregate_peaks_col}{aggregate_peaks_col}

\item{method}{A character string indicating the method to fit the model.
* \code{'glm'} - Generalized Liner Model with \code{\link[stats]{glm}}.
* \code{'glmnet'}, \code{'cv.glmnet'} - Regularized Generalized Liner Model with \code{\link[glmnet]{glmnet}}.
* \code{'brms'} - Bayesian Regression Models using \code{\link[brms]{brms-package}}.
* \code{'xgb'} - Gradient Boosting Regression using \code{\link[xgboost]{xgboost}}.
* \code{'bagging_ridge'} - Bagging Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
* \code{'bayesian_ridge'} - Bayesian Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.}

\item{interaction_term}{The interaction term to use in the model between TF and binding site.
* \code{'+'} for additive interaction.
* \code{':'} for 'multiplicative' interaction.
* \code{'*'} for crossing interaction, i.e. additive AND 'multiplicative'.
For more info, see \code{\link[stats]{formula}}}

\item{adjust_method}{Method for adjusting p-values.}

\item{scale}{Logical. Whether to z-transform the expression and accessibility matrices.}

\item{verbose}{Logical. Display messages. Set verbose to '2' to print errors for all model fits.}
}
\value{
A CSNObject object.
}
\description{
Fit models for gene expression
}
