% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-models.R
\name{fit_model}
\alias{fit_model}
\title{Fit (regularized) generalized linear model}
\usage{
fit_model(
  formula,
  data,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  family = gaussian,
  alpha = 1,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} with a symbolic description
of the model to be fitted.}

\item{data}{A \code{data.frame} containing the variables in the model.}

\item{method}{A character string indicating the method to fit the model.
* \code{'glm'} - Generalized Liner Model with \code{\link[stats]{glm}}.
* \code{'glmnet'}, \code{'cv.glmnet'} - Regularized Generalized Liner Model with \code{\link[glmnet]{glmnet}}.
* \code{'brms'} - Bayesian Regression Models using \code{\link[brms]{brms-package}}.
* \code{'xgb'} - Gradient Boosting Regression using \code{\link[xgboost]{xgboost}}.
* \code{'bagging_ridge'} - Bagging Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.
* \code{'bayesian_ridge'} - Bayesian Ridge Regression using scikit-learn via \link[reticulate]{reticulate}.}

\item{family}{A description of the error distribution and link function to be used in the model.
See \code{\link[stats]{family}} for mode details.}

\item{alpha}{The elasticnet mixing parameter. See \code{\link[glmnet]{glmnet}} for details.}

\item{...}{Other parameters for the model fitting function.}
}
\value{
A list with two data frames: \code{gof} contains goodness of fit measures of the fit and
\code{coefs} contains the fitted coefficients.
}
\description{
Fit (regularized) generalized linear model
}
