% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_srm}
\alias{fit_srm}
\title{Fit a sparse regression model}
\usage{
fit_srm(formula, data, ...)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{...}{Additional parameters passed to the underlying sparse regression function.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{gof}} - Goodness of fit measures
\item \emph{\code{coefs}} - Fitted coefficients with sparse structure
}
}
\description{
Fits a sparse regression model using specialized algorithms for high-dimensional data.
}
