% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_susie}
\alias{fit_susie}
\title{Fit a SuSiE regression model}
\usage{
fit_susie(formula, data, ...)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{...}{Additional parameters passed to \emph{\code{susieR::susie}}.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{metrics}} - Goodness of fit measures
\item \emph{\code{coefficients}} - Fitted coefficients
}
}
\description{
Fits a Sum of Single Effects (SuSiE) regression model using the susieR package.
}
