% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-grn.R
\name{format_coefs}
\alias{format_coefs}
\title{Format network coefficients}
\usage{
format_coefs(coefs, term = ":", adjust_method = "fdr")
}
\arguments{
\item{coefs}{A data frame with coefficients}

\item{term}{term}

\item{adjust_method}{adjust_method}
}
\value{
A data frame.
}
\description{
Format network coefficients
}
