% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{generate.aggregated.data}
\alias{generate.aggregated.data}
\title{Create aggregated data for a certain cluster}
\usage{
generate.aggregated.data(
  object,
  cell_coord,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{cell_coord}{similarity matrix or dimiension reductions.}

\item{k_neigh}{Number of cells to aggregate per cluster.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary}

\item{max_overlap}{The maximum overlapping ratio of two clusters.}

\item{seed}{Random seed}

\item{verbose}{Logical, should warning and info messages be printed?}
}
\value{
Aggregated data.
}
\description{
Function to generate aggregated inputs of a cetrain cluster. \code{generate.aggregated.data}
takes as input sparse data. This function will aggregate binary accessibility scores (or gene expression)
per cell cluster, if they do not overlap any existing cluster with more than 50% cells.
}
