% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate.cre.gene.links}
\alias{generate.cre.gene.links}
\title{Extract cre-gene relations of markers}
\usage{
generate.cre.gene.links(weight_table, markers, filter = FALSE)
}
\arguments{
\item{weight_table}{The result of \code{inferCSN()}}

\item{markers}{A dataframe data with marker genes and cluster information,
such as the results of \code{Seurat::FindAllMarkers(object)}.}

\item{filter}{filter}
}
\description{
Extract cre-gene relations of markers
}
