% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate.links}
\alias{generate.links}
\title{generate links}
\usage{
generate.links(
  cre_gene_list,
  cre_tf_list,
  promoter_cre_gene_list,
  promoter_cre_tf_list,
  clusters
)
}
\arguments{
\item{cre_gene_list}{A list of CRE-gene}

\item{cre_tf_list}{A list of CRE-TF}

\item{promoter_cre_gene_list}{A list of CRE-gene about promoters}

\item{promoter_cre_tf_list}{A list of CRE-TF about promoters}

\item{clusters}{cluster information}
}
\value{
TF-G dataframe
}
\description{
generate links
}
