% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate.nodes}
\alias{generate.nodes}
\title{generate Node character}
\usage{
generate.nodes(network_links, markers)
}
\arguments{
\item{network_links}{Links obtained by generate.links}

\item{markers}{Dataframe with marker and cluster information}
}
\value{
Node character dataframe
}
\description{
generate Node character
}
