% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate.peak.tf.links}
\alias{generate.peak.tf.links}
\title{Identify TFs enriched in cre of focus markers}
\usage{
generate.peak.tf.links(peaks_bed_list, species, genome, markers)
}
\arguments{
\item{peaks_bed_list}{A list of peaks bed file of each cluster}

\item{species}{Species used to detect TF}

\item{genome}{For example: BSgenome.Hsapiens.UCSC.hg38}

\item{markers}{two column dataframe data with marker genes and cluster information}
}
\description{
Identify TFs enriched in cre of focus markers
}
