% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_CRE}
\alias{generate_CRE}
\title{extract CREs of markers}
\usage{
generate_CRE(cre_gene_list, promoter_cre_gene_list, da_peaks_list = NULL)
}
\arguments{
\item{cre_gene_list}{a list of CRE-Gene relationships}

\item{promoter_cre_gene_list}{a list of Promoter-Gene relationships}

\item{da_peaks_list}{a list of DA of each cluster}
}
\description{
extract CREs of markers
}
