% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate_CRE_Gene_links}
\alias{generate_CRE_Gene_links}
\title{extract CREs-gene relations of markers}
\usage{
generate_CRE_Gene_links(weight_table, markers, filter = FALSE)
}
\arguments{
\item{weight_table}{dataframe of the result of inferCSN}

\item{markers}{two column dataframe data with marker genes and cluster information}
}
\description{
extract CREs-gene relations of markers
}
