% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_get.embedding.R
\name{get.dimensional}
\alias{get.dimensional}
\alias{get.dimensional.default}
\alias{get.dimensional.Seurat}
\title{Get dimensional information}
\usage{
get.dimensional(object, ...)

\method{get.dimensional}{default}(object, cores = 1, seed = 1, ...)

\method{get.dimensional}{Seurat}(object, ...)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{cores}{Using n threads for \code{\link[uwot]{umap}}.}

\item{seed}{Set a random seed, default sets to 1.}
}
\value{
Dimensional information
}
\description{
Get dimensional information
}
\examples{
library(inferCSN)
data("example_matrix")
dimensional_information <- get.dimensional(example_matrix)
}
