% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_get.embedding.R
\name{get.embedding}
\alias{get.embedding}
\alias{get.embedding.default}
\alias{get.embedding.Seurat}
\title{Get embedding information}
\usage{
get.embedding(object, ...)

\method{get.embedding}{default}(object, cores = 1, ...)

\method{get.embedding}{Seurat}(object, ...)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{cores}{Using n threads for \code{\link[uwot]{umap}}.}
}
\value{
Embedding information
}
\description{
Get embedding information
}
\examples{
library(inferCSN)
data("example_matrix")
embedding_information <- get.embedding(example_matrix)
}
