% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_get.pseudotime.R
\name{get.pseudotime}
\alias{get.pseudotime}
\alias{get.pseudotime.default}
\alias{get.pseudotime.Seurat}
\title{Get pseudotime information}
\usage{
get.pseudotime(object, ...)

\method{get.pseudotime}{default}(
  object,
  meta_data = NULL,
  embeddings = NULL,
  cluster_by = "cluster",
  cores = 1,
  seed = 1,
  start_cluster = NULL,
  end_cluster = NULL,
  verbose = TRUE,
  ...
)

\method{get.pseudotime}{Seurat}(
  object,
  assay = "RNA",
  cluster_by = "cluster",
  slot = "data",
  start_cluster = NULL,
  end_cluster = NULL,
  ...
)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{meta_data}{meta data}

\item{embeddings}{embeddings information}

\item{cluster_by}{choose column used for `slingshot`}

\item{cores}{CPU cores used for umap}

\item{seed}{random seed for umap}

\item{start_cluster}{start_cluster}

\item{end_cluster}{end_cluster}

\item{verbose}{verbose}

\item{assay}{assay}

\item{slot}{slot}
}
\value{
Dimensional information

A list with a matrix and new meta data

Seurat object
}
\description{
Get pseudotime information
}
\examples{
data("example_matrix")
result <- get.pseudotime(example_matrix)
}
