% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.network.R
\name{get.tf.gene}
\alias{get.tf.gene}
\title{Obtain TF and Gene relationships}
\usage{
get.tf.gene(cre_gene_list, cre_tf_list, clusters)
}
\arguments{
\item{cre_gene_list}{A list of CRE-gene}

\item{cre_tf_list}{A list of CRE-TF}

\item{clusters}{cluster information}
}
\value{
TF-G dataframe
}
\description{
Obtain TF and Gene relationships
}
