% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attribute.R
\name{get_attribute}
\alias{get_attribute}
\alias{get_attribute,CSNObject-method}
\title{Get any attribute from a CSNObject object}
\usage{
get_attribute(object, ...)

\S4method{get_attribute}{CSNObject}(
  object,
  celltypes = NULL,
  active_network = NULL,
  attribute = c("genes", "tfs", "peaks", "regulators", "targets", "celltypes", "cells",
    "modules", "coefficients"),
  ...
)
}
\arguments{
\item{object}{A CSNObject object}

\item{...}{Additional arguments}

\item{celltypes}{A character vector specifying the celltypes to get attributes for.
If \code{NULL}, all celltypes are returned.}

\item{active_network}{A character string specifying the active network to get attributes for.}

\item{attribute}{A character string specifying the attribute to get.
This can take any of the following choices:
\describe{
\item{genes}{The original gene without any filtering.}
\item{tfs}{The original transcription factors without any filtering.}
\item{peaks}{The original peaks without any filtering.}
\item{regulators}{The regulators of the network after inference CSN.}
\item{targets}{The targets of the network after inference CSN.}
\item{cells}{The cells of the network.}
\item{modules}{The modules of the network.}
\item{coefficients}{The coefficients of the network.}
}}
}
\value{
A character vector of regulatory genes
}
\description{
Get any attribute from a CSNObject object
}
