% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod-get_embedding.R
\name{get_embedding}
\alias{get_embedding}
\alias{get_embedding,matrix-method}
\alias{get_embedding,Seurat-method}
\title{Get dimensional information}
\usage{
get_embedding(object, ...)

\S4method{get_embedding}{matrix}(object, cores = 1, seed = 1, dims = 10, ...)

\S4method{get_embedding}{Seurat}(object, reduction = "umap", dims = 2, ...)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{cores}{Using n threads for \code{\link[uwot]{umap}}.}

\item{seed}{Set a random seed, default sets to 1.}

\item{dims}{The number of dimensions to use.}

\item{reduction}{The reduction to use.}
}
\value{
Embedding
}
\description{
Get dimensional information
}
\examples{
\dontrun{
data("example_matrix")
dimensional_information <- get_embedding(example_matrix)
}
}
