% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{get_umap}
\alias{get_umap}
\title{Compute UMAP embedding}
\usage{
get_umap(x, n_pcs = 30, ...)
}
\arguments{
\item{x}{data}

\item{n_pcs}{n_pcs}

\item{...}{other parameters}
}
\value{
umap
}
\description{
Compute UMAP embedding
}
