% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{GetAssay.CSNObject}
\alias{GetAssay.CSNObject}
\alias{GetAssay}
\title{Get Seurat assay from CSNObject}
\usage{
\method{GetAssay}{CSNObject}(object, assay = NULL, ...)

GetAssay(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{assay}{assay}

\item{...}{Arguments for other methods}
}
\description{
Get Seurat assay from CSNObject
}
