% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{GetAssaySummary}
\alias{GetAssaySummary}
\alias{GetAssaySummary,Seurat-method}
\alias{GetAssaySummary,CSNObject-method}
\title{Get summary of seurat assay}
\usage{
GetAssaySummary(object, ...)

\S4method{GetAssaySummary}{Seurat}(object, group_name, assay = NULL, verbose = TRUE, ...)

\S4method{GetAssaySummary}{CSNObject}(object, group_name, assay = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{group_name}{group_name}

\item{assay}{assay}

\item{verbose}{verbose}
}
\description{
Get summary of seurat assay

Get summary of seurat assay from CSNObject
}
