% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{GetGRN.CSNObject}
\alias{GetGRN.CSNObject}
\alias{GetGRN}
\title{Get network data}
\usage{
\method{GetGRN}{CSNObject}(object, ...)

GetGRN(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network data
}
