% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{gof.CSNObject}
\alias{gof.CSNObject}
\alias{gof.RegulatoryNetwork}
\alias{gof.Network}
\alias{gof}
\title{Get goodness-of-fit info}
\usage{
\method{gof}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{gof}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{gof}{Network}(object, network = DefaultNetwork(object), ...)

gof(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{network}{network}

\item{...}{Arguments for other methods}
}
\description{
Get goodness-of-fit info
}
