% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.heatmap.R
\name{hm_dyn_clust}
\alias{hm_dyn_clust}
\title{plots results of findDynGenes}
\usage{
hm_dyn_clust(
  matrix,
  dynRes,
  geneAnn,
  row_cols,
  limits = c(0, 10),
  toScale = FALSE,
  fontsize_row = 4
)
}
\arguments{
\item{matrix}{expression matrix}

\item{dynRes}{dynRes}

\item{geneAnn}{geneAnn}

\item{row_cols}{row_cols}

\item{limits}{limits}

\item{toScale}{toScale}

\item{fontsize_row}{fontsize_row
'}
}
\value{
heatmaps
}
\description{
plots results of findDynGenes
}
