% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hm_dyn_epoch}
\alias{hm_dyn_epoch}
\title{heatmap}
\usage{
hm_dyn_epoch(
  matrix,
  epochRes,
  row_cols,
  limits = c(0, 10),
  toScale = FALSE,
  fontsize_row = 4
)
}
\arguments{
\item{matrix}{expression matrix}

\item{epochRes}{result of running findDynGenes}

\item{row_cols}{row_cols}

\item{limits}{limits}

\item{toScale}{toScale}

\item{fontsize_row}{fontsize_row
'}
}
\value{
heatmaps
}
\description{
heatmap
}
