% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.heatmap.R
\name{ig_convertMedium}
\alias{ig_convertMedium}
\title{change igraph attributes so that it is suitable for plotting a medium network}
\usage{
ig_convertMedium(iG, rCol = "#74a9cf", tCol = "#66C2A5", vScale = 1)
}
\arguments{
\item{iG}{in which regulators and targets are already specified}

\item{rCol}{default color for regulator nodes}

\item{tCol}{default color for target nodes}

\item{vScale}{vScale}
}
\description{
change igraph attributes so that it is suitable for plotting a medium network
}
