% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ig_convertTFs}
\alias{ig_convertTFs}
\title{change igraph attributes so that it is suitable for plotting a network of only regs}
\usage{
ig_convertTFs(
  iG,
  rCol = "#F46D43",
  tCol = "#66C2A5",
  vScale = 1,
  exponent = 2.6
)
}
\arguments{
\item{iG}{in which regulators and targets are already specified}

\item{rCol}{default color for regulator nodes}

\item{tCol}{default color for target nodes}

\item{vScale}{vScale}

\item{exponent}{exponent}
}
\description{
change igraph attributes so that it is suitable for plotting a network of only regs
}
