% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.heatmap.R
\name{ig_NiceGraph}
\alias{ig_NiceGraph}
\title{returns a pretty graph given a grnTab and expression data}
\usage{
ig_NiceGraph(myG, grnTab, vLabels = "Regulator")
}
\arguments{
\item{myG}{result of running induced.subgraph(graphX, Vgenes), and ig_convertSmall}

\item{grnTab}{grnTab}

\item{vLabels}{vLabels}
}
\value{
iGraph object
}
\description{
returns a pretty graph given a grnTab and expression data
}
