% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ig_tabToIgraph}
\alias{ig_tabToIgraph}
\title{convert a table to an igraph}
\usage{
ig_tabToIgraph(grnTab, simplify = FALSE, directed = FALSE, weights = TRUE)
}
\arguments{
\item{grnTab}{table of regulator, regulator, maybe zscores, maybe correlations}

\item{simplify}{false}

\item{directed}{FALSE,}

\item{weights}{TRUE}
}
\value{
iGraph object
}
\description{
This adds an nEnts vertex attribute to count the number of entities in the sub-net
}
