% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod-infer_vector.R
\name{infer_vector}
\alias{infer_vector}
\alias{infer_vector,list-method}
\alias{infer_vector,matrix-method}
\alias{infer_vector,Seurat-method}
\title{infer VECTOR}
\usage{
infer_vector(object, ...)

\S4method{infer_vector}{list}(object, plot = TRUE, ...)

\S4method{infer_vector}{matrix}(object, filter = FALSE, plot = TRUE, ...)

\S4method{infer_vector}{Seurat}(
  object,
  filter = FALSE,
  plot = TRUE,
  reduction = "umap",
  dims = 2,
  ...
)
}
\arguments{
\item{object}{The information of pca and UMAP.}

\item{...}{Arguments for other methods}

\item{plot}{Logical value.}

\item{filter}{Logical value, default set to \code{FALSE},
will filter cells though pseudotime.}

\item{reduction}{The reduction method.}

\item{dims}{The reduction to use.}
}
\value{
vector
}
\description{
infer VECTOR
}
\examples{
\dontrun{
data("example_matrix")
dimensional_information <- get_embedding(example_matrix)
vector_result <- infer_vector(dimensional_information)
}
\dontrun{
data("example_matrix")
vector_result <- infer_vector(
  example_matrix,
  plot = FALSE
)
}
}
