% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\docType{package}
\name{inferCSN-package}
\alias{inferCSN-package}
\title{inferCSN: infer cell-type-specific gene regulatory network}
\source{
\url{https://github.com/mengxu98/inferCSNplus}
}
\description{
An R package for infer cell-type-specific gene regulatory network from single-cell omics data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/inferCSNplus/}
  \item Report bugs at \url{https://github.com/mengxu98/inferCSNplus/issues}
}

}
\author{
Meng xu (Maintainer), \email{mengxu98@qq.com}
}
