% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{inferCSN}
\alias{inferCSN}
\title{Inferring Cell-Specific Gene Regulatory Network}
\usage{
inferCSN(object, ...)
}
\arguments{
\item{object}{The object for inferCSN}

\item{...}{Arguments for other methods}
}
\value{
A data table of gene-gene regulatory relationship
}
\description{
Inferring Cell-Specific Gene Regulatory Network
}
\examples{
library(inferCSN)
data("exampleMatrix")
weightDT <- inferCSN(exampleMatrix, verbose = TRUE)
head(weightDT)

weightDT <- inferCSN(exampleMatrix, verbose = TRUE, cores = 2)
head(weightDT)

}
