% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_VECTOR.R
\name{inferVECTOR}
\alias{inferVECTOR}
\alias{inferVECTOR.default}
\alias{inferVECTOR.VECTOR_dimension}
\alias{inferVECTOR.matrix}
\alias{inferVECTOR.Seurat}
\title{infer VECTOR}
\usage{
inferVECTOR(object, ...)

\method{inferVECTOR}{default}(object, plot = TRUE, ...)

\method{inferVECTOR}{VECTOR_dimension}(object, plot = TRUE, ...)

\method{inferVECTOR}{matrix}(object, filter = TRUE, plot = TRUE, ...)

\method{inferVECTOR}{Seurat}(object, filter = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{The information of PCA and UMAP.}

\item{...}{Arguments for other methods}

\item{plot}{Logical value.}

\item{filter}{Logical value, default set to \code{TRUE},
will filter cells though pseudotime.}
}
\value{
Dimensional information

vector result
}
\description{
infer VECTOR
}
\examples{
data("example_matrix")
dimensional_information <- inferVECTOR(example_matrix)
library(inferCSN)
data("example_matrix")
dimensional_information <- get.dimensional(example_matrix)
vector_result <- inferVECTOR(dimensional_information)
data("example_matrix")
vector_result <- inferVECTOR(
  example_matrix,
  plot = FALSE
)
}
