% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{LayerData}
\alias{LayerData}
\alias{LayerData,CSNObject-method}
\title{Get layer data from CSNObject}
\usage{
LayerData(object, ...)

\S4method{LayerData}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get layer data from CSNObject
}
