% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{Modules-class}
\alias{Modules-class}
\title{The Modules class}
\description{
The Modules object stores the TF modules extracted from the inferred network..
}
\section{Slots}{

\describe{
\item{\code{meta}}{A dataframe with meta data about the modules.}

\item{\code{features}}{A named list with a set of fetures (genes/peaks) for each module.}

\item{\code{params}}{A named list with module selection parameters.}
}}

