% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-objects.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\alias{Network}
\title{The Network class}
\description{
The Network object stores the inferred network itself, information about the fitting
process as well as graph representations of the network.
}
\section{Slots}{

\describe{
\item{\code{modules}}{A list TF modules.}

\item{\code{features}}{A named list containing the transcription factors and
target genes included in the network.}

\item{\code{fit}}{A dataframe with goodness of fit measures.}

\item{\code{coefs}}{A dataframe with the fitted coefficients.}

\item{\code{graphs}}{Graphical representations of the inferred network.}

\item{\code{params}}{A named list with GRN inference parameters.}
}}

