% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{NetworkFeatures}
\alias{NetworkFeatures}
\alias{NetworkFeatures.CSNObject}
\alias{NetworkFeatures.RegulatoryNetwork}
\title{Get network features}
\usage{
NetworkFeatures(object, ...)

\method{NetworkFeatures}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{NetworkFeatures}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}
}
\description{
Get network features
}
