% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{NetworkRegions}
\alias{NetworkRegions}
\alias{NetworkRegions.CSNObject}
\alias{NetworkRegions.RegulatoryNetwork}
\title{Get network regions}
\usage{
NetworkRegions(object, ...)

\method{NetworkRegions}{CSNObject}(object, ...)

\method{NetworkRegions}{RegulatoryNetwork}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network regions
}
