% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.object.R
\name{normalize.seurat.object}
\alias{normalize.seurat.object}
\title{normalize.seurat.object}
\usage{
normalize.seurat.object(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Normalized seurat object
}
\description{
normalize.seurat.object
}
