% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{order_genes}
\alias{order_genes}
\title{Function that orders genes based on peak expression}
\usage{
order_genes(matrix, meta_data, pseudotime_column, smooth = TRUE)
}
\arguments{
\item{matrix}{expression matrix}

\item{meta_data}{sample table}

\item{pseudotime_column}{column in sample table containing pseudotime annotation}

\item{smooth}{whether or not to smooth expression across pseudotime}
}
\value{
ordered genes
}
\description{
Function that orders genes based on peak expression
}
