% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{Params}
\alias{Params}
\alias{Params.CSNObject}
\alias{Params.RegulatoryNetwork}
\title{Get GRN inference parameters}
\usage{
Params(object, ...)

\method{Params}{CSNObject}(object, ...)

\method{Params}{RegulatoryNetwork}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get GRN inference parameters
}
