% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.calling.R
\name{peak.calling}
\alias{peak.calling}
\title{peak.calling}
\usage{
peak.calling(object, fragments = NULL, macs2_path = NULL, verbose = FALSE)
}
\arguments{
\item{object}{An expression matrix, cells by genes.}

\item{fragments}{fragments}

\item{macs2_path}{path to macs2}

\item{verbose}{Print detailed information.}
}
\value{
seurat object
}
\description{
peak.calling
}
