% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_functions.R, R/generics.R
\name{peaks.filter.default}
\alias{peaks.filter.default}
\alias{peaks.filter.Seurat}
\alias{peaks.filter}
\title{peaks.filter}
\usage{
peaks.filter.default(object, filter_value = 0.9)

\method{peaks.filter}{Seurat}(object, filter_value = 0.9)

peaks.filter(object, ...)
}
\arguments{
\item{object}{The object for inferCSN}

\item{filter_value}{The value to filter peaks}

\item{...}{Arguments for other methods}
}
\value{
Returns a object
}
\description{
peaks.filter
}
