% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.overlapping.coordinates.R
\name{peaks.split}
\alias{peaks.split}
\title{peaks.split}
\usage{
peaks.split(inp)
}
\arguments{
\item{inp}{parameter}
}
\value{
seq
}
\description{
This function from:\url{https://github.com/cole-trapnell-lab/cicero-release/blob/master/R/utils.R#L507}
}
