% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{Perturbation-class}
\alias{Perturbation-class}
\title{The Perturbation class}
\description{
The Perturbation object stores the original expression matrix,
 perturbed expression matrix, TF states, prediction results,
 and embedding results.
}
\section{Slots}{

\describe{
\item{\code{original}}{A matrix.}

\item{\code{perturbed}}{A matrix.}

\item{\code{tf_states}}{A dataframe.}

\item{\code{prediction}}{A list.}

\item{\code{embedding}}{A matrix.}

\item{\code{params}}{A list.}
}}

