% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.connections}
\alias{plot.connections}
\title{Plot links of inferCSN on focused genes}
\usage{
\method{plot}{connections}(
  weight_table,
  gene_anno = NULL,
  marker,
  cutoff = NULL,
  upstream = 250000,
  downstream = 250000
)
}
\arguments{
\item{weight_table}{inferCSN result}

\item{gene_anno}{the genome annotation}

\item{marker}{the focused gene}

\item{cutoff}{the cutoff of importance scores.
CREs are links with importance scores higher than cutoff}

\item{upstream}{the number of distance upstream to TSS}

\item{downstream}{the number of distance downstream to TSS}
}
\value{
ggplot2 object
}
\description{
Plot links of inferCSN on focused genes
}
