% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{plot_gene_rank}
\alias{plot_gene_rank}
\alias{plot_gene_rank,Network-method}
\alias{plot_gene_rank,data.frame-method}
\title{Plot gene ranks and network properties}
\usage{
plot_gene_rank(object, ...)

\S4method{plot_gene_rank}{Network}(
  object,
  method = c("page_rank", "degree_distribution"),
  weight_cutoff = 0.1,
  compare_random = TRUE,
  ...
)

\S4method{plot_gene_rank}{data.frame}(
  object,
  method = c("page_rank", "degree_distribution"),
  weight_cutoff = 0.1,
  compare_random = TRUE
)
}
\arguments{
\item{object}{Network object}

\item{...}{Other params}

\item{method}{Character, ranking method: "page_rank" or "degree_distribution"}

\item{weight_cutoff}{Numeric, threshold for edge weight filtering}

\item{compare_random}{Logical, whether to compare with randomized network}
}
\value{
Combined ggplot object
}
\description{
Plot gene ranks and network properties
}
