% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.analysis.R
\name{plot_heatmap_by_treatment}
\alias{plot_heatmap_by_treatment}
\title{Useful plotting function to plot heatmap with pre-split matrix}
\usage{
plot_heatmap_by_treatment(
  expList,
  meta_data,
  pseudotime_column = "latent_time",
  toScale = T,
  limits = c(0, 5),
  smooth = TRUE,
  anno_colors = NA,
  fontSize = 8
)
}
\arguments{
\item{expList}{list of expression matrices}

\item{meta_data}{sample table}

\item{pseudotime_column}{column in sample table containing pseudotime annotation}

\item{toScale}{whether or not to scale expression}

\item{limits}{limits on expression}

\item{smooth}{whether or not to smooth expression across pseudotime for cleaner plotting}

\item{anno_colors}{annotation colors}

\item{fontSize}{heatmap font size}
}
\value{
pheatmap
}
\description{
Useful plotting function to plot heatmap with pre-split matrix
}
