% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(
  data,
  method = "loess",
  group_colors = c("#e1181f", "#255830", "#006699"),
  titleColor = "#006699",
  title = NULL,
  xTitle = NULL,
  yTitle = NULL,
  legendTitle = NULL,
  legend = "bottom"
)
}
\arguments{
\item{data}{data}

\item{method}{method}

\item{group_colors}{group_colors}

\item{titleColor}{titleColor}

\item{title}{title}

\item{xTitle}{xTitle}

\item{yTitle}{yTitle}

\item{legendTitle}{legendTitle}

\item{legend}{legend}
}
\value{
ggplot object
}
\description{
plot_scatter
}
\examples{
data("example_matrix")
test_data <- data.frame(
  example_matrix[, c(1,11)],
  c = c(rep("c1", 2000), rep("c2", 2000), rep("c3", 1000))
)
plot_scatter(test_data)
}
