% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/utils_plots.R
\name{plot_tf_network}
\alias{plot_tf_network}
\alias{plot_tf_network.CSNObject}
\title{Plot sub-network centered around one TF.}
\usage{
plot_tf_network(object, ...)

\method{plot_tf_network}{CSNObject}(
  object,
  tf,
  network = DefaultNetwork(object),
  graph = "module_graph",
  circular = TRUE,
  edge_width = 0.2,
  edge_color = c(`-1` = "darkgrey", `1` = "orange"),
  node_size = 3,
  text_size = 10,
  label_nodes = c("tfs", "all", "none"),
  color_edges = TRUE,
  ...
)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{tf}{The transcription factor to center around.}

\item{network}{Name of the network to use.}

\item{graph}{Name of the graph.}

\item{circular}{Logical. Layout tree in circular layout.}

\item{edge_width}{Edge width.}

\item{edge_color}{Edge color.}

\item{node_size}{Node size.}

\item{text_size}{Font size for labels.}

\item{label_nodes}{String, indicating what to label.
* \code{'tfs'} - Label all TFs.
* \code{'all'} - Label all genes.
* \code{'none'} - Label nothing (except the root TF).}

\item{color_edges}{Logical, whether to color edges by direction.}
}
\value{
A CSNObject object.
}
\description{
Plot sub-network centered around one TF.
}
