% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.heatmap.R
\name{plot_top_features}
\alias{plot_top_features}
\title{quick plot of top regulators in dynamic networks}
\usage{
plot_top_features(
  network_list,
  gene_ranks = NULL,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  targets_num = NULL,
  network_order = NULL,
  method = "weight"
)
}
\arguments{
\item{network_list}{the result of running epochGRN}

\item{gene_ranks}{the result of running compute_pagerank}

\item{regulators}{regulators}

\item{targets}{targets}

\item{regulators_num}{number of top regulators to plot}

\item{targets_num}{number of top targets to plot for each regulator}

\item{network_order}{which epochs or transitions to plot}

\item{method}{Choose method to rank regulators for plot.
Defaule set to `weight`, mean choose regulators rely weights in network infer by \link[inferCSN]{inferCSN}.
Also can choose `page_rank`, then will ranking regulators rely the result of \link[igraph]{page_rank}.}
}
\value{
plot
}
\description{
quick plot of top regulators in dynamic networks
}
