% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{pre_pseudotime_matrix}
\alias{pre_pseudotime_matrix}
\title{pre_pseudotime_matrix}
\usage{
pre_pseudotime_matrix(matrix, gene_list = NULL)
}
\arguments{
\item{matrix}{matrix}

\item{gene_list}{A vector of gene names}
}
\value{
A smoothed pseudotime matrix for the given gene list
}
\description{
Calculate and return a smoothed pseudotime matrix for the given gene list
}
