% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{print.Network}
\alias{print.Network}
\alias{print.Modules}
\alias{print.Regions}
\title{Print Network objects}
\usage{
\method{print}{Network}(x, ...)

\method{print}{Modules}(x, ...)

\method{print}{Regions}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{other parameters}
}
\description{
Print Network objects

Print Modules objects

Print Regions objects
}
