% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.overlapping.coordinates.R
\name{ranges.for.coords}
\alias{ranges.for.coords}
\title{Construct GRanges objects from coordinate strings}
\usage{
ranges.for.coords(coord_strings, meta_data_df = NULL, with_names = FALSE)
}
\arguments{
\item{coord_strings}{A list of coordinate strings (in the form "chr1:500000-1000000").}

\item{meta_data_df}{A data frame with any meta data columns you want
included with the ranges. Must be in the same order as coord_strings.}

\item{with_names}{logical, should meta data include coordinate string (field coord_string)?}
}
\value{
GRanges object of the input strings
}
\description{
This function from:\url{https://github.com/cole-trapnell-lab/cicero-release/blob/master/R/utils.R#L90}
}
\details{
Coordinate strings consist of three pieces of information:
  chromosome, start, and stop. These pieces of information can be separated
  by the characters ":", "_", or "-". Commas will be removed, not used as
  separators (ex: "chr18:8,575,097-8,839,855" is ok).
}
\examples{
ran1 <- ranges.for.coords("chr1:2039-30239", with_names = TRUE)
ran1

ran2 <- ranges.for.coords(
  c("chr1:2049-203902", "chrX:489249-1389389"),
  meta_data_df = data.frame(dat = c("1", "X"))
)
ran2

ran3 <- ranges.for.coords(
  c("chr1:2049-203902", "chrX:489249-1389389"),
  with_names = TRUE,
  meta_data_df = data.frame(
    dat = c("1", "X"),
    stringsAsFactors = FALSE
  )
)
ran3
}
\seealso{
\code{\link[GenomicRanges]{GRanges-class}}
}
