% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-objects.R
\docType{class}
\name{RegulatoryNetwork-class}
\alias{RegulatoryNetwork-class}
\alias{RegulatoryNetwork}
\title{The RegularotyNetwork class}
\description{
The RegularotyNetwork object is the core data structure in Pando.
It stores all data necessary for network inference and analysis
that is not provided by Seurat.
}
\section{Slots}{

\describe{
\item{\code{regions}}{A \code{\link{Regions}} object containing information about
the genomic regions included in the network.}

\item{\code{networks}}{A \code{\link{Network}} object containing the inferred regulatory
network and information about the model fit.}

\item{\code{params}}{A list storing parameters for GRN inference.}

\item{\code{active_network}}{A string indicating the active network.}
}}

