% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rough_hierarchy}
\alias{rough_hierarchy}
\title{rough_hierarchy}
\usage{
rough_hierarchy(network_table, abs_weight = TRUE, directed = TRUE)
}
\arguments{
\item{network_table}{The weight data table of network.}

\item{abs_weight}{Logical value, default is \emph{\code{TRUE}},
whether to perform absolute value on weights,
and when set \code{abs_weight} to \emph{\code{TRUE}},
the output of weight table will create a new column named \code{Interaction}.}

\item{directed}{Whether the network is directed.}
}
\value{
list
}
\description{
returns rough roots in the network, rough roots selected as those connected to most number of nodes
}
\examples{
data("example_matrix")
network_table <- inferCSN(example_matrix)
rough_hierarchy(network_table)
}
