% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_perturbation.R
\name{SetNetwork}
\alias{SetNetwork}
\title{Set Network object}
\usage{
SetNetwork(object, network, name = NULL)
}
\arguments{
\item{object}{CSNObject}

\item{network}{Network object to set}

\item{name}{Character, name of the network (optional)}
}
\value{
Updated CSNObject
}
\description{
Set Network object
}
