% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_csn_data}
\alias{simulate_csn_data}
\title{Simulate data for testing inferCSN}
\usage{
simulate_csn_data(
  n_cells = 300,
  n_genes = 100,
  n_tfs = 50,
  n_peaks = 1000,
  n_celltypes = 3,
  sparsity = 0.9,
  atac_sparsity = 0.9
)
}
\arguments{
\item{n_cells}{Number of cells to simulate}

\item{n_genes}{Number of genes to simulate}

\item{n_tfs}{Number of transcription factors}

\item{n_peaks}{Number of ATAC peaks to simulate}

\item{n_celltypes}{Number of cell types to simulate}

\item{sparsity}{Sparsity level of the expression matrix (0-1)}

\item{atac_sparsity}{Sparsity level of the ATAC matrix (0-1)}
}
\value{
A list containing simulated data
}
\description{
Simulate data for testing inferCSN
}
