% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.regression.R
\name{sparse.regression}
\alias{sparse.regression}
\title{Sparse regression model}
\usage{
sparse.regression(
  X,
  y,
  crossValidation = FALSE,
  seed = 1,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{X}{The data matrix}

\item{y}{The response vector}
}
\value{
The coefficients
}
\description{
Sparse regression model
}
