% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network.R
\name{split_epochs_by_group}
\alias{split_epochs_by_group}
\title{Splits data into epochs manually}
\usage{
split_epochs_by_group(dynamic_object, assignment)
}
\arguments{
\item{dynamic_object}{result of running findDynGenes or compileDynGenes}

\item{assignment}{a list of vectors where names(assignment) are epoch names, and vectors contain groups belonging to corresponding epoch}
}
\value{
updated dynamic_object with epoch column included in dynamic_object$cells
}
\description{
Splits data into epochs given group assignment
}
