% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.regression.R
\name{sub.inferCSN}
\alias{sub.inferCSN}
\title{Sparse regression model for single gene}
\usage{
sub.inferCSN(
  regulatorsMatrix,
  targetsMatrix,
  target = NULL,
  crossValidation = FALSE,
  seed = 1,
  penalty = "L0",
  algorithm = "CD",
  maxSuppSize = NULL,
  nFolds = 10,
  kFolds = NULL,
  rThreshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{regulatorsMatrix}{Regulators matrix}

\item{targetsMatrix}{Targets matrix}

\item{target}{Target genes}
}
\value{
The weight data table of sub-network
}
\description{
Sparse regression model for single gene
}
