% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.analysis.R
\name{subnets}
\alias{subnets}
\title{Function to assign nodes to communities via Louvain clustering}
\usage{
subnets(df, tfs, tfonly = TRUE)
}
\arguments{
\item{df}{dataframe containing a static network}

\item{tfs}{TFs}

\item{tfonly}{if TRUE will limit network to TFs only}
}
\value{
dataframe containing community assignments for each gene
}
\description{
Function to assign nodes to communities via Louvain clustering
}
