% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{summary_csn}
\alias{summary_csn}
\alias{summary_csn.CSNObject}
\title{Get network data}
\usage{
summary_csn(object, ...)

\method{summary_csn}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network data
}
