% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{transcript.mode}
\alias{transcript.mode}
\title{use gaussian kernel to calculate the mode of transcript counts}
\usage{
transcript.mode(x, breaks = "Sturges")
}
\arguments{
\item{x}{log tranformed relative expression}

\item{breaks}{control parameter}
}
\description{
use gaussian kernel to calculate the mode of transcript counts
}
