% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{VariableFeatures.CSNObject}
\alias{VariableFeatures.CSNObject}
\alias{VariableFeatures}
\title{Get variable features from CSNObject}
\usage{
\method{VariableFeatures}{CSNObject}(object, ...)

VariableFeatures(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get variable features from CSNObject
}
