% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{VECTOR-class}
\alias{VECTOR-class}
\title{The Modules class}
\description{
The Modules object stores the TF modules extracted from the inferred network.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{Filtered matrix.}

\item{\code{pseduotime}}{Pseduotime of cells}
}}

