% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{vector.autoCenter}
\alias{vector.autoCenter}
\title{Title}
\usage{
vector.autoCenter(OUT, UP = 0.9, plot = TRUE)
}
\arguments{
\item{OUT}{OUT}

\item{UP}{UP}

\item{plot}{plot}
}
\value{
output
}
\description{
Title
}
