% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector.buildGrid}
\alias{vector.buildGrid}
\title{Title}
\usage{
vector.buildGrid(VEC, N = 30, plot = TRUE, COL = "grey70")
}
\arguments{
\item{VEC}{VEC}

\item{N}{N}

\item{plot}{plot}

\item{COL}{COL}
}
\value{
output
}
\description{
Title
}
