% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector.buildNet}
\alias{vector.buildNet}
\title{Title}
\usage{
vector.buildNet(OUT, CUT = 1, plot = TRUE, COL = "grey70")
}
\arguments{
\item{OUT}{OUT}

\item{CUT}{CUT}

\item{plot}{plot}

\item{COL}{COL}
}
\value{
output
}
\description{
Title
}
