% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector.drawArrow}
\alias{vector.drawArrow}
\title{Title}
\usage{
vector.drawArrow(
  OUT,
  P = 0.9,
  plot = TRUE,
  COL = "grey70",
  OL = 1.5,
  arrow_length2 = 70,
  CEX = 0.5,
  arrow_width = 1,
  BD = TRUE,
  arrow_color = "grey30",
  plot.SUMMIT = TRUE
)
}
\arguments{
\item{OUT}{OUT}

\item{P}{P}

\item{plot}{plot}

\item{COL}{COL}

\item{OL}{OL}

\item{arrow_length2}{arrow_length2}

\item{CEX}{CEX}

\item{arrow_width}{arrow_width}

\item{BD}{BD}

\item{arrow_color}{arrow_color}

\item{plot.SUMMIT}{plot.SUMMIT}
}
\value{
output
}
\description{
Title
}
