% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector.main}
\alias{vector.main}
\title{vector.main}
\usage{
vector.main(embedding_information, plot = TRUE)
}
\arguments{
\item{embedding_information}{embedding_information}

\item{plot}{plot}
}
\value{
vector_result
}
\description{
vector.main
}
\examples{
library(inferCSN)
data("example_matrix")
embedding_information <- get.embedding(example_matrix)
vector_result <- vector.main(embedding_information)
}
