% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector_auto_center}
\alias{vector_auto_center}
\title{Title}
\usage{
vector_auto_center(out, up_value = 0.9, plot = TRUE)
}
\arguments{
\item{out}{out}

\item{up_value}{up_value}

\item{plot}{plot}
}
\value{
output
}
\description{
Title
}
