% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector_build_grid}
\alias{vector_build_grid}
\title{Title}
\usage{
vector_build_grid(vector, N = 30, plot = TRUE, colors = "grey70")
}
\arguments{
\item{vector}{vector}

\item{N}{N}

\item{plot}{plot}

\item{colors}{colors}
}
\value{
output
}
\description{
Title
}
