% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector_build_net}
\alias{vector_build_net}
\title{Title}
\usage{
vector_build_net(out, CUT = 1, plot = TRUE, colors = "grey70")
}
\arguments{
\item{out}{out}

\item{CUT}{CUT}

\item{plot}{plot}

\item{colors}{colors}
}
\value{
output
}
\description{
Title
}
