% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector_build_value}
\alias{vector_build_value}
\title{Title}
\usage{
vector_build_value(out, umap, plot = TRUE)
}
\arguments{
\item{out}{out}

\item{umap}{umap}

\item{plot}{plot}
}
\value{
output
}
\description{
Title
}
