% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vector.R
\name{vector_draw_arrow}
\alias{vector_draw_arrow}
\title{Title}
\usage{
vector_draw_arrow(
  out,
  P = 0.9,
  plot = TRUE,
  colors = "grey70",
  OL = 1.5,
  arrow_length2 = 70,
  cex_value = 0.5,
  arrow_width = 1,
  BD = TRUE,
  arrow_color = "grey30",
  plot.SUMMIT = TRUE
)
}
\arguments{
\item{out}{out}

\item{P}{P}

\item{plot}{plot}

\item{colors}{colors}

\item{OL}{OL}

\item{arrow_length2}{arrow_length2}

\item{cex_value}{cex_value}

\item{arrow_width}{arrow_width}

\item{BD}{BD}

\item{arrow_color}{arrow_color}

\item{plot.SUMMIT}{plot.SUMMIT}
}
\value{
output
}
\description{
Title
}
