% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_rank.R
\name{calculate_gene_rank}
\alias{calculate_gene_rank}
\alias{calculate_gene_rank,Network-method}
\alias{calculate_gene_rank,data.frame-method}
\alias{calculate_degree_distribution}
\alias{calculate_degree_distribution,Network-method}
\alias{calculate_degree_distribution,data.frame-method}
\title{Calculate gene rank}
\usage{
calculate_gene_rank(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)

\S4method{calculate_gene_rank}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)

\S4method{calculate_gene_rank}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = FALSE,
  method = c("page_rank", "degree_distribution"),
  ...
)

calculate_degree_distribution(
  object,
  regulators = NULL,
  targets = NULL,
  directed = TRUE
)

\S4method{calculate_degree_distribution}{Network}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = TRUE
)

\S4method{calculate_degree_distribution}{data.frame}(
  object,
  regulators = NULL,
  targets = NULL,
  directed = TRUE
)
}
\arguments{
\item{object}{Network object}

\item{regulators}{Character vector, regulators to include}

\item{targets}{Character vector, targets to include}

\item{directed}{Logical, whether the network is directed}

\item{method}{Character, ranking method: "page_rank" or "degree_distribution"}

\item{...}{Additional arguments}
}
\value{
Data frame with gene ranks
}
\description{
Calculate gene rank
}
