% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{coef.CSNObject}
\alias{coef.CSNObject}
\alias{coef.Network}
\title{Get fitted coefficients}
\usage{
\method{coef}{CSNObject}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\method{coef}{Network}(object, ...)
}
\arguments{
\item{object}{The csn object}

\item{network}{network}

\item{celltypes}{cell types to analyze, NULL for all cell types}

\item{...}{other parameters}
}
\description{
Get fitted coefficients
}
