% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{coef.CSNObject}
\alias{coef.CSNObject}
\alias{coef.RegulatoryNetwork}
\alias{coef.Network}
\title{Get fitted coefficients}
\usage{
\method{coef}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{coef}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{coef}{Network}(object, ...)
}
\arguments{
\item{object}{The grn object}

\item{network}{network}

\item{...}{other parameters}
}
\description{
Get fitted coefficients
}
