% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{DefaultNetwork}
\alias{DefaultNetwork}
\alias{DefaultNetwork.CSNObject}
\alias{DefaultNetwork.RegulatoryNetwork}
\title{Get active network}
\usage{
DefaultNetwork(object, ...)

\method{DefaultNetwork}{CSNObject}(object, ...)

\method{DefaultNetwork}{RegulatoryNetwork}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get active network
}
