% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{DefaultNetwork.CSNObject}
\alias{DefaultNetwork.CSNObject}
\alias{DefaultNetwork.RegulatoryNetwork}
\alias{DefaultNetwork}
\title{Get active network}
\usage{
\method{DefaultNetwork}{CSNObject}(object, ...)

\method{DefaultNetwork}{RegulatoryNetwork}(object, ...)

DefaultNetwork(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get active network
}
