% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_dynamic.genes.R
\name{dynamic.genes}
\alias{dynamic.genes}
\alias{dynamic.genes.default}
\alias{dynamic.genes.Seurat}
\title{Get dynamic genes}
\usage{
dynamic.genes(object, ...)

\method{dynamic.genes}{default}(object, pseudotime = NULL, fdr_threshold = 0.05, cores = 1, ...)

\method{dynamic.genes}{Seurat}(
  object,
  fdr_threshold = 0.05,
  cluster = NULL,
  cluster_column = NULL,
  pseudotime_column = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{object}{Expression object.}

\item{...}{Arguments for other methods}

\item{pseudotime}{Pseudotime}

\item{fdr_threshold}{Threshold of fdr.}

\item{cores}{cores}

\item{cluster}{cluster}

\item{cluster_column}{idents}

\item{pseudotime_column}{pseudotime_column}

\item{slot}{slot used}
}
\value{
Dimensional information

Gene list
}
\description{
Get dynamic genes

Select dynamic genes
}
\examples{
data("example_matrix")
vector_result <- inferVECTOR(example_matrix)
dynamic.genes(
  object = t(vector_result$matrix),
  pseudotime = vector_result$pseudotime[, 2]
)
}
