% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod-dynamic_genes.R
\name{dynamic_genes}
\alias{dynamic_genes}
\alias{dynamic_genes,matrix-method}
\alias{dynamic_genes,Seurat-method}
\title{Get dynamic genes}
\usage{
dynamic_genes(object, ...)

\S4method{dynamic_genes}{matrix}(
  object,
  pseudotime = NULL,
  cores = 1,
  verbose = TRUE,
  adjust_method = "BH",
  ...
)

\S4method{dynamic_genes}{Seurat}(
  object,
  pseudotime_column,
  cores = 1,
  verbose = TRUE,
  adjust_method = "BH",
  layer = "data",
  ...
)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{pseudotime}{Pseudotime of cells, the length and order should be the same as the expression matrix columns.}

\item{cores}{The number of cores to use for parallelization with \code{\link[foreach]{foreach}}, default is \emph{\code{1}}.}

\item{verbose}{Logical value, default is \emph{\code{TRUE}}, whether to print progress messages.}

\item{adjust_method}{The method used to calculate adjust P-value.}

\item{pseudotime_column}{The column name in meta_data annotating pseudotime or latent time.}

\item{layer}{The layer used in Seurat object.}
}
\value{
A new object with dynamic genes
}
\description{
Get dynamic genes
}
\examples{
data("example_matrix")
data("example_meta_data")
dynamic_genes(
  object = example_matrix,
  pseudotime = example_meta_data$pseudotime
)

\dontrun{
vector_result <- infer_vector(example_matrix)
dynamic_genes(
  object = vector_result$matrix,
  pseudotime = vector_result$pseudotime[, 2]
)
}
}
