% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{dynamic_shortest_path_multiple}
\alias{dynamic_shortest_path_multiple}
\title{Function to return shortest path from multiple TFs to multiple targets in a dynamic network}
\usage{
dynamic_shortest_path_multiple(grn, from, to, weight_column = "weight")
}
\arguments{
\item{grn}{a dyanmic network}

\item{from}{the starting TFs}

\item{to}{the end TFs}

\item{weight_column}{column name in network_table with edge weights that will be converted to distances}
}
\value{
dataframe with shortest path, distance, normalized distance, and action
}
\description{
Function to return shortest path from multiple TFs to multiple targets in a dynamic network
}
