% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{expression_ksmooth}
\alias{expression_ksmooth}
\title{expression_ksmooth}
\usage{
expression_ksmooth(
  matrix,
  meta_data,
  pseudotime_column = "pseudotime",
  bandwith = 0.25
)
}
\arguments{
\item{matrix}{expression matrix.}

\item{meta_data}{meta_data includes "cell_name", "pseudotime", "group".}

\item{pseudotime_column}{pseudotime_column.}

\item{bandwith}{bandwith for kernel smoother.}
}
\value{
smoothed matrix
}
\description{
smooths expression across cells in path.
}
