% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_cvsrm}
\alias{fit_cvsrm}
\title{Cross-validation for sparse regression models}
\usage{
fit_cvsrm(formula, data, verbose = TRUE, ...)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{verbose}{If `TRUE`, show warning messages.}

\item{...}{Additional parameters passed to the underlying sparse regression function.}
}
\value{
A list containing two data frames:
* `metrics` - Goodness of fit measures
* `coefficients` - Fitted coefficients with sparse structure
}
\description{
Fits a sparse regression model with cross-validation.
}
\examples{
data("example_matrix")
df <- as.data.frame(example_matrix)
fit_cvsrm(g1 ~ ., data = df)
}
