% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(
  formula,
  data,
  method = c("srm", "glm", "glmnet", "cv.glmnet", "brms", "xgb", "bagging_ridge",
    "bayesian_ridge"),
  family = gaussian,
  alpha = 1,
  ...
)
}
\arguments{
\item{formula}{An object of class \emph{\code{formula}} with a symbolic description
of the model to be fitted.}

\item{data}{A \emph{\code{data.frame}} containing the variables in the model.}

\item{method}{A character string indicating the method to fit the model.
This can take any of the following choices:
\itemize{
\item \emph{\code{srm}} - Sparse Regression Model
\item \emph{\code{glm}} - Generalized Linear Model using \emph{\code{stats::glm}}
\item \emph{\code{glmnet}}, \emph{\code{cv.glmnet}} - Regularized GLM using \emph{\code{glmnet::glmnet}}
\item \emph{\code{brms}} - Bayesian Regression using \emph{\code{brms}} package
\item \emph{\code{xgb}} - Gradient Boosting using \emph{\code{xgboost}}
\item \emph{\code{bagging_ridge}} - Bagging Ridge Regression via scikit-learn
\item \emph{\code{bayesian_ridge}} - Bayesian Ridge Regression via scikit-learn
}}

\item{family}{A description of the error distribution and link function.
See \emph{\code{stats::family}} for details.}

\item{alpha}{The elasticnet mixing parameter, between 0 and 1.
See \emph{\code{glmnet::glmnet}} for details.}

\item{...}{Additional parameters passed to the underlying model fitting function.}
}
\value{
A list containing two data frames:
\itemize{
\item \emph{\code{gof}} - Goodness of fit measures
\item \emph{\code{coefs}} - Fitted coefficients
}
}
\description{
Fits various types of regression models including generalized linear models,
regularized models, Bayesian models, and machine learning approaches.
}
