% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_2.R
\name{format_coefs}
\alias{format_coefs}
\title{Format network coefficients}
\usage{
format_coefs(coefficients, variable = ":", adjust_method = "fdr")
}
\arguments{
\item{coefficients}{A data frame with coefficients}

\item{variable}{variable}

\item{adjust_method}{adjust_method}
}
\value{
A data frame.
}
\description{
Format network coefficients
}
