% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethod.R
\name{get_attribute}
\alias{get_attribute}
\alias{get_attribute,CSNObject-method}
\title{Get regulatory genes}
\usage{
get_attribute(object, ...)

\S4method{get_attribute}{CSNObject}(
  object,
  celltypes = NULL,
  active_network = NULL,
  attribute = c("genes", "tfs", "peaks", "regulators", "targets", "celltypes", "cells",
    "modules", "coefficients"),
  ...
)
}
\arguments{
\item{object}{A CSNObject object}

\item{...}{Additional arguments}

\item{celltypes}{A character vector specifying the celltypes to get attributes for.
If \code{NULL}, all celltypes are returned.}

\item{active_network}{A character string specifying the active network to get attributes for.}

\item{attribute}{A character string specifying the attribute to get.
This can take any of the following choices:
\itemize{
\item \emph{\code{genes}} - Gene names
\item \emph{\code{tfs}} - Transcription factors
\item \emph{\code{peaks}} - Peaks
\item \emph{\code{regulators}} - Regulators
\item \emph{\code{targets}} - Targets
\item \emph{\code{cells}} - Cells
\item \emph{\code{modules}} - Modules
\item \emph{\code{coefficients}} - Coefficients
}}
}
\value{
A character vector of regulatory genes
}
\description{
Get regulatory genes
}
