% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod-get_pseudotime.R
\name{get_pseudotime}
\alias{get_pseudotime}
\alias{get_pseudotime.default}
\alias{get_pseudotime.Seurat}
\title{Get pseudotime information}
\usage{
get_pseudotime(object, ...)

\method{get_pseudotime}{default}(
  object,
  meta_data = NULL,
  embeddings = NULL,
  cluster_by = "cluster",
  cores = 1,
  seed = 1,
  start_cluster = NULL,
  end_cluster = NULL,
  verbose = TRUE,
  ...
)

\method{get_pseudotime}{Seurat}(
  object,
  assay = "RNA",
  cluster_by = "cluster",
  slot = "data",
  reduction = "umap",
  start_cluster = NULL,
  end_cluster = NULL,
  ...
)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{meta_data}{Input meta data.}

\item{embeddings}{Embeddings information}

\item{cluster_by}{The column used for `slingshot`}

\item{cores}{CPU cores used for umap}

\item{seed}{random seed for umap}

\item{start_cluster}{start_cluster}

\item{end_cluster}{end_cluster}

\item{verbose}{Logical, whether to print log messages.}

\item{assay}{The assay used for `slingshot`.}

\item{slot}{The slot used for `slingshot`.}

\item{reduction}{The reduction used for `slingshot`.}
}
\value{
Dimensional information

A list with a matrix and new meta data

Seurat object
}
\description{
Get pseudotime information
}
\examples{
\dontrun{
data("example_matrix")
result <- get_pseudotime(example_matrix)
}
}
