% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/get_pseudotime.R
\name{get_pseudotime}
\alias{get_pseudotime}
\alias{get_pseudotime.default}
\alias{get_pseudotime.Seurat}
\title{Get pseudotime information}
\usage{
get_pseudotime(object, ...)

\method{get_pseudotime}{default}(
  object,
  meta_data = NULL,
  embeddings = NULL,
  cluster_column = "cluster",
  cores = 1,
  seed = 1,
  start_cluster = NULL,
  end_cluster = NULL,
  verbose = TRUE,
  ...
)

\method{get_pseudotime}{Seurat}(
  object,
  assay = "RNA",
  layer = "data",
  cluster_column = "cluster",
  reduction = "umap",
  start_cluster = NULL,
  end_cluster = NULL,
  ...
)
}
\arguments{
\item{object}{The input data, a matrix with cells/samples by genes/features or a seurat object.}

\item{...}{Arguments for other methods}

\item{meta_data}{Input meta data.}

\item{embeddings}{Embeddings information}

\item{cluster_column}{The column used for \code{\link[slingshot]{slingshot}}.}

\item{cores}{The number of cores to use for parallelization with \code{\link[foreach]{foreach}}, default is \emph{\code{1}}.}

\item{seed}{The random seed for cross-validation, default is \emph{\code{1}}.}

\item{start_cluster}{The start cluster.}

\item{end_cluster}{The end cluster.}

\item{verbose}{Logical value, default is \emph{\code{TRUE}}, whether to print progress messages.}

\item{assay}{The assay used for \code{\link[slingshot]{slingshot}}.}

\item{layer}{The layer used for \code{\link[slingshot]{slingshot}}.}

\item{reduction}{The reduction used for \code{\link[slingshot]{slingshot}}.}
}
\value{
Pseudotime information corresponding to cells
}
\description{
Get pseudotime information
}
\examples{
\dontrun{
data("example_matrix")
head(get_pseudotime(example_matrix))
}
}
