% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay,CSNObject-method}
\title{Get seurat assay}
\usage{
GetAssay(object, ...)

\S4method{GetAssay}{CSNObject}(object, assay = NULL, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{assay}{assay}
}
\description{
Get seurat assay

Get Seurat assay from CSNObject
}
