% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay.CSNObject}
\title{Get Seurat assay from CSNObject}
\usage{
GetAssay(object, ...)

\method{GetAssay}{CSNObject}(object, assay = NULL, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}

\item{assay}{assay}
}
\description{
Get Seurat assay from CSNObject
}
