% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{GetAssaySummary.Seurat}
\alias{GetAssaySummary.Seurat}
\alias{GetAssaySummary.CSNObject}
\alias{GetAssaySummary}
\title{Get summary of seurat assay}
\usage{
\method{GetAssaySummary}{Seurat}(object, group_name, assay = NULL, verbose = TRUE, ...)

\method{GetAssaySummary}{CSNObject}(object, group_name, assay = NULL, verbose = TRUE, ...)

GetAssaySummary(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{group_name}{group_name}

\item{assay}{assay}

\item{verbose}{verbose}

\item{...}{Arguments for other methods}
}
\description{
Get summary of seurat assay

Get summary of seurat assay from CSNObject
}
