% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{GetGRN}
\alias{GetGRN}
\alias{GetGRN.CSNObject}
\title{Get network data}
\usage{
GetGRN(object, ...)

\method{GetGRN}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network data
}
