% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{GetNetwork.CSNObject}
\alias{GetNetwork.CSNObject}
\alias{GetNetwork.RegulatoryNetwork}
\alias{GetNetwork}
\title{Get network}
\usage{
\method{GetNetwork}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{GetNetwork}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

GetNetwork(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{network}{network}

\item{...}{Arguments for other methods}
}
\description{
Get network

Get network
}
