% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{GetNetwork}
\alias{GetNetwork}
\alias{GetNetwork.CSNObject}
\alias{GetNetwork.RegulatoryNetwork}
\title{Get network}
\usage{
GetNetwork(object, ...)

\method{GetNetwork}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{GetNetwork}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}
}
\description{
Get network
}
