% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{gof}
\alias{gof}
\alias{gof.CSNObject}
\alias{gof.RegulatoryNetwork}
\alias{gof.Network}
\title{Get goodness-of-fit info}
\usage{
gof(object, ...)

\method{gof}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{gof}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{gof}{Network}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}
}
\description{
Get goodness-of-fit info
}
