% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCSN-package.R
\docType{package}
\name{inferCSN-package}
\alias{inferCSN-package}
\title{Inferring cell-type-specific gene regulatory networks from single-cell omics data}
\source{
\url{https://github.com/mengxu98/inferCSN}
}
\description{
An package for inferring cell-type-specific gene regulatory networks from single-cell omics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mengxu98.github.io/inferCSNplus/}
  \item Report bugs at \url{https://github.com/mengxu98/inferCSNplus/issues}
}

}
\author{
Meng xu (Maintainer), \email{mengxu98@qq.com}
}
