% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod_inferCSN.R
\name{inferCSN}
\alias{inferCSN}
\alias{inferCSN.default}
\alias{inferCSN.data.frame}
\alias{inferCSN.Seurat}
\title{Inferring Cell-Specific Gene Regulatory Network}
\usage{
inferCSN(object, ...)

\method{inferCSN}{default}(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  k_folds = NULL,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  verbose = FALSE,
  cores = 1,
  ...
)

\method{inferCSN}{data.frame}(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  k_folds = NULL,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  verbose = FALSE,
  cores = 1,
  ...
)

\method{inferCSN}{Seurat}(
  object,
  penalty = "L0",
  algorithm = "CD",
  cross_validation = FALSE,
  seed = 1,
  n_folds = 10,
  k_folds = NULL,
  r_threshold = 0,
  regulators = NULL,
  targets = NULL,
  regulators_num = NULL,
  verbose = FALSE,
  cores = 1,
  aggregate = TRUE,
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction_name = NULL,
  size_factor_normalize = FALSE,
  genome_info = NULL,
  high_corr_cutoff = NULL,
  low_corr_cutoff = NULL,
  rescued = TRUE,
  ...
)
}
\arguments{
\item{object}{An expression matrix, cells by genes}

\item{...}{Additional arguments.}

\item{penalty}{The type of regularization.
This can take either one of the following choices: "L0" and "L0L2".
For high-dimensional and sparse data, such as single-cell sequencing data, "L0L2" is more effective.}

\item{algorithm}{The type of algorithm used to minimize the objective function.
Currently "CD" and "CDPSI" are supported.
The CDPSI algorithm may yield better results, but it also increases running time.}

\item{cross_validation}{Check whether cross validation is used.}

\item{seed}{The seed used in randomly shuffling the data for cross-validation.}

\item{n_folds}{The number of folds for cross-validation.}

\item{k_folds}{The number of folds for sample split.}

\item{r_threshold}{r_threshold.}

\item{regulators}{Regulator genes.}

\item{targets}{Target genes.}

\item{regulators_num}{The number of non-zore coef, this value will affect the final performance.
The maximum support size at which to terminate the regularization path.
Recommend setting this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small portion of non-zeros.}

\item{verbose}{Print detailed information.}

\item{cores}{CPU cores.}

\item{aggregate}{Logical, whether to aggregate the data.}

\item{k_neigh}{Number of cells to be aggregated per cluster.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary}

\item{max_overlap}{The maximum overlapping ratio of two clusters.}

\item{reduction_name}{The reduction name of extracting the cell coordinates used for aggregating.}

\item{size_factor_normalize}{Logical, should accessibility values be normalized by size factor.}

\item{genome_info}{Genome information.}

\item{high_corr_cutoff}{atacbinary}

\item{low_corr_cutoff}{atacbinary}

\item{rescued}{Logical}
}
\value{
A data table of gene-gene regulatory relationship
}
\description{
Inferring Cell-Specific Gene Regulatory Network
}
\examples{
library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix, verbose = TRUE)
head(weight_table)

weight_table <- inferCSN(example_matrix, verbose = TRUE, cores = 2)
head(weight_table)

\dontrun{
data("promoter_regions_hg38")
seurat_object <- inferCSN(seurat_object, genome_info = promoter_regions_hg38)
}
}
