% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-regions.R
\name{initiate_object}
\alias{initiate_object}
\alias{initiate_object.Seurat}
\alias{initiate_object.CSNObject}
\title{Initiate the \code{RegulatoryNetwork} object.}
\usage{
initiate_object(object, ...)

\method{initiate_object}{Seurat}(
  object,
  regions = NULL,
  peak_assay = "peaks",
  rna_assay = "RNA",
  exclude_exons = TRUE,
  ...
)

\method{initiate_object}{CSNObject}(
  object,
  regions = NULL,
  peak_assay = "peaks",
  rna_assay = "RNA",
  exclude_exons = TRUE,
  ...
)
}
\arguments{
\item{object}{An object.}

\item{...}{Arguments for other methods}

\item{regions}{Candidate regions to consider for binding site inference.
If \code{NULL}, all peaks regions are considered.}

\item{peak_assay}{A character vector indicating the name of the chromatin
accessibility assay in the \code{Seurat} object.}

\item{rna_assay}{A character vector indicating the name of the gene expression
assay in the \code{Seurat} object.}

\item{exclude_exons}{Logical. Whether to consider exons for binding site inference.}
}
\value{
A CSNObject object containing a RegulatoryNetwork object.
}
\description{
Initiate the \code{RegulatoryNetwork} object.

Initiate the \code{RegulatoryNetwork} object.
}
