% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{LayerData}
\alias{LayerData}
\alias{LayerData.CSNObject}
\title{Get layer data from CSNObject}
\usage{
LayerData(object, ...)

\method{LayerData}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{other parameters}
}
\description{
Get layer data from CSNObject
}
