% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{LayerData.CSNObject}
\alias{LayerData.CSNObject}
\alias{LayerData}
\title{Get layer data from CSNObject}
\usage{
\method{LayerData}{CSNObject}(object, ...)

LayerData(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get layer data from CSNObject
}
