% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{NetworkFeatures.CSNObject}
\alias{NetworkFeatures.CSNObject}
\alias{NetworkFeatures.RegulatoryNetwork}
\alias{NetworkFeatures}
\title{Get network features}
\usage{
\method{NetworkFeatures}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{NetworkFeatures}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

NetworkFeatures(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{network}{network}

\item{...}{Arguments for other methods}
}
\description{
Get network features
}
