% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{NetworkGraph.CSNObject}
\alias{NetworkGraph.CSNObject}
\alias{NetworkGraph.RegulatoryNetwork}
\alias{NetworkGraph.Network}
\alias{NetworkGraph}
\title{Get network parameters}
\usage{
\method{NetworkGraph}{CSNObject}(
  object,
  network = DefaultNetwork(object),
  graph = "module_graph",
  ...
)

\method{NetworkGraph}{RegulatoryNetwork}(
  object,
  network = DefaultNetwork(object),
  graph = "module_graph",
  ...
)

\method{NetworkGraph}{Network}(object, graph = "module_graph", ...)

NetworkGraph(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{network}{network}

\item{graph}{graph}

\item{...}{Arguments for other methods}
}
\description{
Get network parameters
}
