% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{NetworkModules.CSNObject}
\alias{NetworkModules.CSNObject}
\alias{NetworkModules.RegulatoryNetwork}
\alias{NetworkModules.Network}
\alias{NetworkModules}
\title{Get TF modules}
\usage{
\method{NetworkModules}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{NetworkModules}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{NetworkModules}{Network}(object, ...)

NetworkModules(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{network}{network}

\item{...}{Arguments for other methods}
}
\description{
Get TF modules
}
