% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{NetworkParams}
\alias{NetworkParams}
\alias{NetworkParams.CSNObject}
\alias{NetworkParams.RegulatoryNetwork}
\alias{NetworkParams.Network}
\title{Get network parameters}
\usage{
NetworkParams(object, ...)

\method{NetworkParams}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{NetworkParams}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{NetworkParams}{Network}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}
}
\description{
Get network parameters
}
