% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{NetworkParams.CSNObject}
\alias{NetworkParams.CSNObject}
\alias{NetworkParams.RegulatoryNetwork}
\alias{NetworkParams.Network}
\alias{NetworkParams}
\title{Get network parameters}
\usage{
\method{NetworkParams}{CSNObject}(object, network = DefaultNetwork(object), ...)

\method{NetworkParams}{RegulatoryNetwork}(object, network = DefaultNetwork(object), ...)

\method{NetworkParams}{Network}(object, ...)

NetworkParams(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{network}{network}

\item{...}{Arguments for other methods}
}
\description{
Get network parameters
}
