% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{NetworkParams}
\alias{NetworkParams}
\alias{NetworkParams.CSNObject}
\alias{NetworkParams.Network}
\title{Get network parameters}
\usage{
NetworkParams(object, ...)

\method{NetworkParams}{CSNObject}(object, network = DefaultNetwork(object), celltypes = NULL, ...)

\method{NetworkParams}{Network}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{network}

\item{celltypes}{cell types to analyze, NULL for all cell types}
}
\description{
Get network parameters
}
