% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{NetworkRegions}
\alias{NetworkRegions}
\alias{NetworkRegions.CSNObject}
\title{Get network regions}
\usage{
NetworkRegions(object, ...)

\method{NetworkRegions}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network regions
}
