% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{NetworkRegions.CSNObject}
\alias{NetworkRegions.CSNObject}
\alias{NetworkRegions.RegulatoryNetwork}
\alias{NetworkRegions}
\title{Get network regions}
\usage{
\method{NetworkRegions}{CSNObject}(object, ...)

\method{NetworkRegions}{RegulatoryNetwork}(object, ...)

NetworkRegions(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network regions
}
