% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{NetworkTFs}
\alias{NetworkTFs}
\alias{NetworkTFs.CSNObject}
\alias{NetworkTFs.RegulatoryNetwork}
\title{Get network TFs}
\usage{
NetworkTFs(object, ...)

\method{NetworkTFs}{CSNObject}(object, ...)

\method{NetworkTFs}{RegulatoryNetwork}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get network TFs
}
