% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setMethod.R
\name{Params}
\alias{Params}
\alias{Params.CSNObject}
\title{Get GRN inference parameters}
\usage{
Params(object, ...)

\method{Params}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get GRN inference parameters
}
