% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R, R/setGenerics.R
\name{Params.CSNObject}
\alias{Params.CSNObject}
\alias{Params.RegulatoryNetwork}
\alias{Params}
\title{Get GRN inference parameters}
\usage{
\method{Params}{CSNObject}(object, ...)

\method{Params}{RegulatoryNetwork}(object, ...)

Params(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get GRN inference parameters
}
