% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.filter.R
\name{peaks.filter}
\alias{peaks.filter}
\title{peaks.filter}
\usage{
peaks.filter(object, filter_value = 0.9)
}
\arguments{
\item{object}{matrix or seurat object}

\item{filter_value}{filter_value}
}
\description{
peaks.filter
}
