% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/utils_plots.R
\name{plot_module_metrics}
\alias{plot_module_metrics}
\alias{plot_module_metrics.CSNObject}
\title{Plot module metrics number of genes, number of peaks and number of TFs per gene.}
\usage{
plot_module_metrics(object, ...)

\method{plot_module_metrics}{CSNObject}(object, network = DefaultNetwork(object), ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}

\item{network}{Name of the network to use.}
}
\value{
A ggplot2 object.
}
\description{
Plot module metrics number of genes, number of peaks and number of TFs per gene.
}
