% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.analysis.R
\name{rough_hierarchy}
\alias{rough_hierarchy}
\title{rough_hierarchy}
\usage{
rough_hierarchy(modnet, abs_weight = TRUE)
}
\arguments{
\item{modnet}{modnet}

\item{abs_weight}{abs_weight}
}
\value{
list
}
\description{
returns rough roots in the network, rough roots selected as those connected to most number of nodes
}
\examples{
library(inferCSN)
data("example_matrix")
weight_table <- inferCSN(example_matrix, verbose = TRUE)
rough_hierarchy(weight_table)
}
