% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{static_shortest_path}
\alias{static_shortest_path}
\title{Function to return shortest path from 1 regulator to 1 target in a static network}
\usage{
static_shortest_path(
  network_table,
  from,
  to,
  weight_column = "weight",
  compare_to_average = FALSE
)
}
\arguments{
\item{network_table}{a static network dataframe}

\item{from}{the starting regulator/gene}

\item{to}{the end regulator/gene}

\item{weight_column}{column name in network_table with edge weights that will be converted to distances}

\item{compare_to_average}{if TRUE will compute normalized against average path length}
}
\value{
shortest path, distance, normalized distance, and action
}
\description{
Function to return shortest path from 1 regulator to 1 target in a static network
}
