% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-generics.R, R/1-objects.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures.CSNObject}
\title{Get variable features from CSNObject}
\usage{
VariableFeatures(object, ...)

\method{VariableFeatures}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a grn object.}

\item{...}{Arguments for other methods}
}
\description{
Get variable features from CSNObject
}
