% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGenerics.R, R/setClass.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures.CSNObject}
\title{Get variable features from CSNObject}
\usage{
VariableFeatures(object, ...)

\method{VariableFeatures}{CSNObject}(object, ...)
}
\arguments{
\item{object}{The input data, a csn object.}

\item{...}{Arguments for other methods}
}
\description{
Get variable features from CSNObject
}
