% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector.autoCenter}
\alias{vector.autoCenter}
\title{Title}
\usage{
vector.autoCenter(OUT, UP = 0.9, SHOW = TRUE)
}
\arguments{
\item{OUT}{OUT}

\item{UP}{UP}

\item{SHOW}{SHOW}
}
\value{
output
}
\description{
Title
}
