% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector.drawArrow}
\alias{vector.drawArrow}
\title{Title}
\usage{
vector.drawArrow(
  OUT,
  P = 0.9,
  SHOW = TRUE,
  COL = "grey70",
  OL = 1.5,
  AL = 60,
  CEX = 0.5,
  AW = 2,
  BD = TRUE,
  AC = "grey20",
  SHOW.SUMMIT = TRUE
)
}
\arguments{
\item{OUT}{OUT}

\item{P}{P}

\item{SHOW}{SHOW}

\item{COL}{COL}

\item{OL}{OL}

\item{AL}{AL}

\item{CEX}{CEX}

\item{AW}{AW}

\item{BD}{BD}

\item{AC}{AC}

\item{SHOW.SUMMIT}{SHOW.SUMMIT}
}
\value{
output
}
\description{
Title
}
