% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{vector.getValue}
\alias{vector.getValue}
\title{Title}
\usage{
vector.getValue(OUT, PCA, plot = TRUE)
}
\arguments{
\item{OUT}{OUT}

\item{PCA}{PCA}

\item{plot}{plot}
}
\value{
output
}
\description{
Title
}
