% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector.main}
\alias{vector.main}
\title{vector.main}
\usage{
vector.main(embedding_information, plot = TRUE)
}
\arguments{
\item{embedding_information}{The information of PCA and UMAP.}

\item{plot}{Logical value.}
}
\value{
vector_result
}
\description{
Raw source code form: \url{https://github.com/jumphone/Vector}.
Author: Feng Zhang.
Paper: Unsupervised Inference of Developmental Directions for Single Cells Using VECTOR,
please reference: \url{https://doi.org/10.1016/j.celrep.2020.108069}.
}
\examples{
library(inferCSN)
data("example_matrix")
embedding_information <- get.embedding(example_matrix)
vector_result <- vector.main(embedding_information)
}
